Split HTTP transport into focused modules
