+++ AGENTS.md
@@ -135,12 +135,12 @@ metadata/update/DNS/TLS inspection modes, and executes requests via `src/http`.
 - Rust formatting code has a central `core::Printer`/`PrinterHandle` and ANSI `Sequence` abstraction; JSON/NDJSON write through the printer directly, formatter/progress/timing/inspection helpers route escape emission through the shared printer, and stderr metadata/error/warning renderers use the same printer for request/response headers and `--inspect-dns`/`--inspect-tls`. Production renderers should expose printer-oriented `*_to(..., &mut core::Printer)` entry points where practical; boolean color wrappers are test-only compatibility helpers.
 - CLI stdout paths that are not already using a stream/pager writer should use `core::write_stdout` or another checked writer instead of `print!`/`println!`, including metadata, completion, gRPC discovery, and WebSocket message output.
 - Rust stdio terminal detection and color/format auto policy are centralized in `src/core.rs`; prefer `core::stdio()`, `core::color_enabled`, and `core::format_enabled` over direct `IsTerminal` checks or local auto-policy helpers.
-- The fetch-owned HTTP transport lives in `src/http/transport.rs`, using hyper-util's pooled client for HTTP/1.1 and HTTP/2, h3/quinn for HTTP/3, and rustls/aws-lc-rs for TLS. Shared network dialing for DNS-aware TCP, HTTP CONNECT, SOCKS5/SOCKS5H, and proxy auth lives in `src/net.rs`; HTTP, WebSocket, DoH, and update code should reuse those helpers instead of duplicating dialer code.
+- The fetch-owned HTTP transport lives under `src/http/transport/`: `client.rs` owns the client API, pooling, TCP/TLS connector, and request helpers; `h3.rs` owns HTTP/3 discovery, cache use, QUIC connections, and race logic; `proxy.rs` owns proxy selection and proxy dialing; `body.rs` owns transport body/response wrappers. `mod.rs` should stay a thin re-export and shared error boundary. Shared network dialing for DNS-aware TCP, HTTP CONNECT, SOCKS5/SOCKS5H, and proxy auth lives in `src/net.rs`; HTTP, WebSocket, DoH, and update code should reuse those helpers instead of duplicating dialer code.
 - Rust error rendering uses rich diagnostics for common CLI/config errors, styling labels, flags/options, invalid values, file paths, and config line context while preserving plain-text `Display` output.
 - Config-backed options are declared in `src/config/mod.rs`'s `config_options!` descriptor table, which generates storage fields and drives parsing, host overlay, CLI source tracking, config application, and CLI/docs parity tests. Add new config-backed options there instead of editing separate option lists.
 - Duplicate host config sections are rejected during config parsing with a line-numbered error that also reports the first section line; do not rely on replacement or merge semantics.
 - Rust `-vvv` output prints config, DNS, TCP, TLS/QUIC, and TTFB debug metadata through the central printer, including color policy and the blank response-header separator before formatted bodies.
-- Rust `--timing` enables DNS pre-resolution timing and transport connection timing so the waterfall includes DNS, TCP, TLS/QUIC, TTFB, and Body phases. The direct transport owns DNS/TCP/TLS/QUIC setup, so keep new timing instrumentation in `src/http/transport.rs` and `src/net.rs`.
+- Rust `--timing` enables DNS pre-resolution timing and transport connection timing so the waterfall includes DNS, TCP, TLS/QUIC, TTFB, and Body phases. The direct transport owns DNS/TCP/TLS/QUIC setup, so keep new timing instrumentation in `src/http/transport/` and `src/net.rs`.
 - `src/main.rs` runs the Tokio runtime on an explicitly larger stack thread, and the top-level app future in `src/app.rs` is heap-pinned before the shutdown-signal `tokio::select!`; do not move it back to `tokio::pin!` or the default process main stack because the combined async request/WebSocket/inspection state can overflow Windows' smaller main-thread stack even for metadata commands.
 - Rust response body paging is controlled by `--pager auto|on|off` or `pager = ...`; `auto` routes terminal stdout through `$PAGER` when set, otherwise falls back to `less -FIRX`, `on` forces the pager, and `off` disables it. `NO_PAGER` disables the default `auto` pager, and `$LESS` is honored by not adding fallback `less` flags when it is set. `$PAGER` is split with POSIX shell-style quoting but launched directly, so shell operators are not interpreted. Image responses and output-file writes bypass the pager.
 - Timeout duration parsing, Go-style duration formatting, elapsed request budgets, connect/DNS timeout caps, and shared `request timed out after ...` errors live in `src/duration.rs`; HTTP, WebSocket, DNS inspection, and TLS inspection paths should reuse `TimeoutBudget` instead of recomputing remaining time locally, and response body deadlines should preserve the original request timeout diagnostic.
