--- src/http/transport.rs
@@ -1,3203 +1,0 @@
-use std::collections::HashMap;
-use std::collections::VecDeque;
-use std::convert::Infallible;
-use std::error::Error as StdError;
-use std::fmt;
-use std::future::{self, Future};
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
-use std::pin::Pin;
-use std::sync::Arc;
-use std::task::{Context, Poll};
-use std::time::Duration;
-
-use base64::Engine;
-use bytes::{Buf, Bytes};
-use futures_util::{Stream, StreamExt, TryStreamExt, stream::FuturesUnordered};
-use http::header::{
-    AUTHORIZATION, CONTENT_LENGTH, COOKIE, HOST, HeaderMap, HeaderValue, PROXY_AUTHORIZATION,
-    SET_COOKIE,
-};
-use http::{Method, Request, Uri, Version};
-use http_body::Frame;
-use http_body_util::{BodyExt, Empty, Full, StreamBody, combinators::UnsyncBoxBody};
-use hyper::body::Incoming;
-use hyper_util::client::legacy::Client as HyperClient;
-use hyper_util::client::legacy::connect::{Connected, Connection};
-use hyper_util::client::proxy::matcher;
-use hyper_util::rt::{TokioExecutor, TokioIo, TokioTimer};
-use quinn::crypto::rustls::QuicClientConfig;
-use rustls::pki_types::ServerName;
-#[cfg(unix)]
-use tokio::net::UnixStream;
-use tokio::sync::Mutex;
-use tokio::task::JoinHandle;
-use tokio_rustls::TlsConnector;
-use tokio_util::io::ReaderStream;
-use tower_service::Service;
-use url::Url;
-
-use crate::cli::HttpVersion;
-use crate::dns::svcb::{HttpsRecordResolver, SvcbRecord};
-use crate::duration::{TimeoutBudget, request_timeout_message};
-use crate::error::FetchError;
-use crate::http::http3_cache::{Http3Cache, Http3CacheCandidate};
-use crate::timing::{DnsTiming, TransportTiming};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ErrorKind {
-    Request,
-    Connect,
-    Timeout,
-    Body,
-}
-
-#[derive(Debug)]
-pub struct Error {
-    kind: ErrorKind,
-    message: String,
-    source: Option<Box<dyn StdError + Send + Sync>>,
-}
-
-impl Error {
-    fn new(kind: ErrorKind, message: impl Into<String>) -> Self {
-        Self {
-            kind,
-            message: message.into(),
-            source: None,
-        }
-    }
-
-    fn with_source(
-        kind: ErrorKind,
-        message: impl Into<String>,
-        source: impl StdError + Send + Sync + 'static,
-    ) -> Self {
-        Self {
-            kind,
-            message: message.into(),
-            source: Some(Box::new(source)),
-        }
-    }
-
-    fn request(message: impl Into<String>) -> Self {
-        Self::new(ErrorKind::Request, message)
-    }
-
-    fn connect(message: impl Into<String>) -> Self {
-        Self::new(ErrorKind::Connect, message)
-    }
-
-    fn timeout(message: impl Into<String>) -> Self {
-        Self::new(ErrorKind::Timeout, message)
-    }
-
-    fn body(message: impl Into<String>) -> Self {
-        Self::new(ErrorKind::Body, message)
-    }
-
-    fn body_source(source: impl StdError + Send + Sync + 'static) -> Self {
-        let message = source.to_string();
-        Self::with_source(ErrorKind::Body, message, source)
-    }
-
-    fn from_fetch(kind: ErrorKind, err: FetchError) -> Self {
-        let message = err.to_string();
-        let kind = if message.starts_with("request timed out after ") {
-            ErrorKind::Timeout
-        } else {
-            kind
-        };
-        Self::new(kind, message)
-    }
-
-    pub(crate) fn is_timeout(&self) -> bool {
-        self.kind == ErrorKind::Timeout
-    }
-
-    pub(crate) fn is_connect(&self) -> bool {
-        self.kind == ErrorKind::Connect
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(&self.message)
-    }
-}
-
-impl StdError for Error {
-    fn source(&self) -> Option<&(dyn StdError + 'static)> {
-        self.source
-            .as_deref()
-            .map(|source| source as &(dyn StdError + 'static))
-    }
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct BodyDeadline {
-    deadline: tokio::time::Instant,
-    timeout_message: String,
-}
-
-impl BodyDeadline {
-    #[cfg(test)]
-    pub(crate) fn new(request_timeout: Duration) -> Self {
-        Self::with_message(request_timeout, request_timeout_message(request_timeout))
-    }
-
-    fn with_message(timeout: Duration, timeout_message: String) -> Self {
-        Self {
-            deadline: tokio::time::Instant::now() + timeout,
-            timeout_message,
-        }
-    }
-
-    fn timeout_error(&self) -> Error {
-        Error::timeout(self.timeout_message.clone())
-    }
-}
-
-pub struct Body {
-    inner: UnsyncBoxBody<Bytes, Error>,
-}
-
-impl Body {
-    pub(crate) fn wrap_stream<S, E>(stream: S) -> Self
-    where
-        S: Stream<Item = Result<Bytes, E>> + Send + 'static,
-        E: fmt::Display + Send + Sync + 'static,
-    {
-        let stream = stream
-            .map_ok(Frame::data)
-            .map_err(|err| Error::body(err.to_string()));
-        Self::boxed(StreamBody::new(stream))
-    }
-
-    fn boxed<B>(body: B) -> Self
-    where
-        B: http_body::Body<Data = Bytes, Error = Error> + Send + 'static,
-    {
-        Self {
-            inner: BodyExt::boxed_unsync(body),
-        }
-    }
-
-    fn map_incoming(body: Incoming) -> Self {
-        Self::boxed(body.map_err(Error::body_source))
-    }
-}
-
-impl From<Bytes> for Body {
-    fn from(bytes: Bytes) -> Self {
-        Self::boxed(Full::new(bytes).map_err(|err| match err {}))
-    }
-}
-
-impl From<Vec<u8>> for Body {
-    fn from(bytes: Vec<u8>) -> Self {
-        Self::from(Bytes::from(bytes))
-    }
-}
-
-impl From<String> for Body {
-    fn from(value: String) -> Self {
-        Self::from(Bytes::from(value))
-    }
-}
-
-impl From<&'static str> for Body {
-    fn from(value: &'static str) -> Self {
-        Self::from(Bytes::from_static(value.as_bytes()))
-    }
-}
-
-impl From<tokio::fs::File> for Body {
-    fn from(file: tokio::fs::File) -> Self {
-        Self::wrap_stream(ReaderStream::new(file))
-    }
-}
-
-impl http_body::Body for Body {
-    type Data = Bytes;
-    type Error = Error;
-
-    fn poll_frame(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
-        Pin::new(&mut self.inner).poll_frame(cx)
-    }
-
-    fn is_end_stream(&self) -> bool {
-        self.inner.is_end_stream()
-    }
-
-    fn size_hint(&self) -> http_body::SizeHint {
-        self.inner.size_hint()
-    }
-}
-
-#[derive(Clone)]
-pub struct Client {
-    config: Arc<ClientConfig>,
-    pooled: HyperClient<TransportConnector, Body>,
-    h3_pool: Arc<Mutex<HashMap<String, H3PooledClient>>>,
-}
-
-type H3SendRequest = h3::client::SendRequest<h3_quinn::OpenStreams, Bytes>;
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct AutoHttp3Config {
-    pub(crate) addrs: Vec<SocketAddr>,
-}
-
-#[derive(Clone)]
-struct H3PooledClient {
-    origin: String,
-    sender: H3SendRequest,
-    remote_addr: SocketAddr,
-}
-
-struct Http3ConnectResult {
-    client: H3PooledClient,
-    timing: TransportTiming,
-}
-
-struct AutoTcpConnection {
-    stream: PooledStream,
-    negotiated_h2: bool,
-    remote_addr: Option<SocketAddr>,
-    timing: TransportTiming,
-}
-
-enum AutoRaceWinner {
-    Http3(Http3ConnectResult),
-    Tcp(AutoTcpConnection),
-}
-
-#[derive(Clone)]
-struct ClientConfig {
-    mode: Option<HttpVersion>,
-    unix_socket: Option<String>,
-    dns_overrides: HashMap<String, Vec<SocketAddr>>,
-    proxies: Vec<Proxy>,
-    tls_config: Option<rustls::ClientConfig>,
-    request_timeout: Option<Duration>,
-    request_timeout_message: Option<String>,
-    connect_timeout: Option<Duration>,
-    session: Option<Arc<crate::session::PersistentCookieStore>>,
-    connection_timing: Option<crate::http::client::ConnectionTiming>,
-    dns_resolution: Option<crate::http::client::DnsResolutionHandle>,
-    dns_server: Option<String>,
-    local_address: Option<IpAddr>,
-    auto_http3: Option<AutoHttp3Config>,
-    auto_http3_discovery: bool,
-    http3_cache: Option<Arc<Http3Cache>>,
-    learn_alt_svc: bool,
-}
-
-pub(crate) struct ClientBuilder {
-    config: ClientConfig,
-}
-
-impl fmt::Debug for ClientBuilder {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("ClientBuilder")
-    }
-}
-
-impl Client {
-    pub(crate) fn builder() -> ClientBuilder {
-        ClientBuilder {
-            config: ClientConfig {
-                mode: None,
-                unix_socket: None,
-                dns_overrides: HashMap::new(),
-                proxies: Vec::new(),
-                tls_config: None,
-                request_timeout: None,
-                request_timeout_message: None,
-                connect_timeout: None,
-                session: None,
-                connection_timing: None,
-                dns_resolution: None,
-                dns_server: None,
-                local_address: None,
-                auto_http3: None,
-                auto_http3_discovery: false,
-                http3_cache: None,
-                learn_alt_svc: false,
-            },
-        }
-    }
-
-    pub(crate) fn request(&self, method: Method, mut url: Url) -> RequestBuilder {
-        let mut headers = HeaderMap::new();
-        if let Some((username, password)) = extract_url_basic_auth(&mut url) {
-            headers.insert(
-                AUTHORIZATION,
-                basic_auth_header_value(&username, password.as_deref()),
-            );
-        }
-        RequestBuilder {
-            client: self.clone(),
-            method,
-            url,
-            headers,
-            body: None,
-            version: None,
-            timeout: self.config.request_timeout,
-            timeout_message: self.config.request_timeout_message.clone(),
-        }
-    }
-
-    pub(crate) fn post(&self, url: Url) -> RequestBuilder {
-        self.request(Method::POST, url)
-    }
-
-    async fn execute(&self, request: RequestBuilder) -> Result<Response, Error> {
-        let version = request
-            .version
-            .or_else(|| self.config.mode.map(version_for_cli));
-        let timeout = TimeoutBudget::new(request.timeout);
-        let body_deadline = request.timeout.map(|timeout| {
-            BodyDeadline::with_message(
-                timeout,
-                request
-                    .timeout_message
-                    .unwrap_or_else(|| request_timeout_message(timeout)),
-            )
-        });
-        timeout
-            .run(self.send(
-                request.method,
-                request.url,
-                request.headers,
-                request.body,
-                version,
-                body_deadline,
-            ))
-            .await
-            .map_err(|err| Error::from_fetch(ErrorKind::Request, err))?
-    }
-
-    async fn send(
-        &self,
-        method: Method,
-        url: Url,
-        mut headers: HeaderMap,
-        body: Option<Body>,
-        version: Option<Version>,
-        body_deadline: Option<BodyDeadline>,
-    ) -> Result<Result<Response, Error>, FetchError> {
-        self.apply_session_cookies(&url, &mut headers);
-        self.apply_proxy_authorization(&url, &mut headers)?;
-        apply_host_header(&url, &mut headers)?;
-        let response = match version {
-            None if (self.config.auto_http3.is_some()
-                || self.config.auto_http3_discovery
-                || self.config.http3_cache.is_some())
-                && url.scheme() == "https" =>
-            {
-                self.send_auto_http3(method, url.clone(), headers, body, body_deadline)
-                    .await
-            }
-            None | Some(Version::HTTP_11 | Version::HTTP_10 | Version::HTTP_2) => {
-                self.send_pooled(method, url.clone(), headers, body, version, body_deadline)
-                    .await
-            }
-            Some(Version::HTTP_3) => {
-                self.send_http3(method, url.clone(), headers, body, body_deadline)
-                    .await
-            }
-            Some(version) => Err(Error::request(format!(
-                "unsupported HTTP version: {version:?}"
-            ))),
-        };
-        if let Ok(response) = &response {
-            self.store_http3_alt_svc(&url, response.headers());
-            self.store_response_cookies(&url, response.headers());
-        }
-        Ok(response)
-    }
-
-    fn apply_session_cookies(&self, url: &Url, headers: &mut HeaderMap) {
-        if headers.contains_key(COOKIE) {
-            return;
-        }
-        let Some(session) = &self.config.session else {
-            return;
-        };
-        if let Some(cookies) = session.cookies(url) {
-            headers.insert(COOKIE, cookies);
-        }
-    }
-
-    fn store_response_cookies(&self, url: &Url, headers: &HeaderMap) {
-        let Some(session) = &self.config.session else {
-            return;
-        };
-        session.set_cookies(&mut headers.get_all(SET_COOKIE).iter(), url);
-    }
-
-    fn store_http3_alt_svc(&self, url: &Url, headers: &HeaderMap) {
-        if !self.config.learn_alt_svc {
-            return;
-        }
-        let Some(cache) = &self.config.http3_cache else {
-            return;
-        };
-        cache.store_alt_svc(url, self.config.dns_server.as_deref(), headers);
-    }
-
-    fn apply_proxy_authorization(
-        &self,
-        url: &Url,
-        headers: &mut HeaderMap,
-    ) -> Result<(), FetchError> {
-        if url.scheme() != "http" || headers.contains_key(PROXY_AUTHORIZATION) {
-            return Ok(());
-        }
-        let Some(proxy) = proxy_for_config(&self.config, url) else {
-            return Ok(());
-        };
-        if !proxy.is_http_proxy() {
-            return Ok(());
-        }
-        let Some(auth) = proxy.basic_auth()? else {
-            return Ok(());
-        };
-        headers.insert(
-            PROXY_AUTHORIZATION,
-            HeaderValue::from_str(&auth).map_err(|err| {
-                FetchError::Message(format!("invalid proxy authorization: {err}"))
-            })?,
-        );
-        Ok(())
-    }
-
-    async fn send_pooled(
-        &self,
-        method: Method,
-        url: Url,
-        headers: HeaderMap,
-        body: Option<Body>,
-        version: Option<Version>,
-        body_deadline: Option<BodyDeadline>,
-    ) -> Result<Response, Error> {
-        let body = body.unwrap_or_else(|| Body::from(Bytes::new()));
-        let request_version = version.unwrap_or(Version::HTTP_11);
-        let request = build_request(method, absolute_uri(&url)?, request_version, headers, body)
-            .map_err(Error::request)?;
-        let response = self
-            .pooled
-            .request(request)
-            .await
-            .map_err(map_pooled_client_error)?;
-        Ok(Response::from_hyper(url, response, body_deadline))
-    }
-
-    async fn send_auto_http3(
-        &self,
-        method: Method,
-        url: Url,
-        headers: HeaderMap,
-        body: Option<Body>,
-        body_deadline: Option<BodyDeadline>,
-    ) -> Result<Response, Error> {
-        match self.race_auto_http3_connection(url.clone()).await? {
-            AutoRaceWinner::Http3(result) => {
-                if let Some(timing) = &self.config.connection_timing {
-                    timing.set(result.timing);
-                }
-                self.store_http3_client(result.client.clone()).await;
-                let (req, body) = build_h3_request(method, &url, headers, body)?;
-                self.send_http3_request(url, req, body, body_deadline, result.client)
-                    .await
-            }
-            AutoRaceWinner::Tcp(connection) => {
-                if let Some(timing) = &self.config.connection_timing {
-                    timing.set(connection.timing);
-                }
-                self.send_tcp_one_shot(method, url, headers, body, body_deadline, connection)
-                    .await
-            }
-        }
-    }
-
-    async fn send_tcp_one_shot(
-        &self,
-        method: Method,
-        url: Url,
-        headers: HeaderMap,
-        body: Option<Body>,
-        body_deadline: Option<BodyDeadline>,
-        connection: AutoTcpConnection,
-    ) -> Result<Response, Error> {
-        let body = body.unwrap_or_else(|| Body::from(Bytes::new()));
-        let version = if connection.negotiated_h2 {
-            Version::HTTP_2
-        } else {
-            Version::HTTP_11
-        };
-        let uri = if connection.negotiated_h2 {
-            absolute_uri(&url)?
-        } else {
-            origin_form_uri(&url)?
-        };
-        let remote_addr = connection.remote_addr;
-        let request = build_request(method, uri, version, headers, body).map_err(Error::request)?;
-        let io = TokioIo::new(connection.stream);
-        let response = if connection.negotiated_h2 {
-            let (mut sender, conn) = hyper::client::conn::http2::Builder::new(TokioExecutor::new())
-                .timer(TokioTimer::new())
-                .handshake(io)
-                .await
-                .map_err(|err| {
-                    Error::with_source(ErrorKind::Connect, format!("http2 handshake: {err}"), err)
-                })?;
-            tokio::spawn(async move {
-                let _ = conn.await;
-            });
-            sender
-                .send_request(request)
-                .await
-                .map_err(|err| Error::with_source(ErrorKind::Request, err.to_string(), err))?
-        } else {
-            let (mut sender, conn) = hyper::client::conn::http1::Builder::new()
-                .handshake(io)
-                .await
-                .map_err(|err| {
-                    Error::with_source(ErrorKind::Connect, format!("http1 handshake: {err}"), err)
-                })?;
-            tokio::spawn(async move {
-                let _ = conn.await;
-            });
-            sender
-                .send_request(request)
-                .await
-                .map_err(|err| Error::with_source(ErrorKind::Request, err.to_string(), err))?
-        };
-        Ok(Response::from_hyper_with_remote(
-            url,
-            response,
-            body_deadline,
-            remote_addr,
-        ))
-    }
-
-    async fn race_auto_http3_connection(&self, url: Url) -> Result<AutoRaceWinner, Error> {
-        if self.config.auto_http3_discovery
-            || (self.config.auto_http3.is_none() && self.config.http3_cache.is_some())
-        {
-            return self.race_dynamic_auto_http3_connection(url).await;
-        }
-        let connect_timeout = TimeoutBudget::new(self.config.connect_timeout);
-        race_primary_fallback(
-            {
-                let client = self.clone();
-                let url = url.clone();
-                move || async move {
-                    client
-                        .connect_auto_http3_client(&url, connect_timeout)
-                        .await
-                        .map(AutoRaceWinner::Http3)
-                }
-            },
-            {
-                let client = self.clone();
-                move || {
-                    let client = client.clone();
-                    let url = url.clone();
-                    async move {
-                        client
-                            .connect_auto_tcp_tls(&url, connect_timeout)
-                            .await
-                            .map(AutoRaceWinner::Tcp)
-                    }
-                }
-            },
-            crate::net::HAPPY_EYEBALLS_FALLBACK_DELAY,
-        )
-        .await
-    }
-
-    async fn race_dynamic_auto_http3_connection(&self, url: Url) -> Result<AutoRaceWinner, Error> {
-        let connect_timeout = TimeoutBudget::new(self.config.connect_timeout);
-        let h3 = {
-            let client = self.clone();
-            let url = url.clone();
-            async move {
-                client
-                    .connect_dynamic_auto_http3_client(&url, connect_timeout)
-                    .await
-                    .map(AutoRaceWinner::Http3)
-            }
-        };
-        let tcp = {
-            let client = self.clone();
-            async move {
-                client
-                    .connect_auto_tcp_tls(&url, connect_timeout)
-                    .await
-                    .map(AutoRaceWinner::Tcp)
-            }
-        };
-        race_started_auto_connections(h3, tcp).await
-    }
-
-    async fn connect_dynamic_auto_http3_client(
-        &self,
-        url: &Url,
-        timeout: TimeoutBudget,
-    ) -> Result<Http3ConnectResult, Error> {
-        let origin = http3_origin(url)?;
-        if let Some(client) = self.h3_pool.lock().await.get(&origin).cloned() {
-            return Ok(Http3ConnectResult {
-                client,
-                timing: TransportTiming::default(),
-            });
-        }
-        let host = url
-            .host_str()
-            .ok_or_else(|| Error::request("URL host is required"))?;
-        let discovery_start = std::time::Instant::now();
-        let cached_candidates = self
-            .config
-            .http3_cache
-            .as_ref()
-            .map(|cache| cache.candidates(url, self.config.dns_server.as_deref()))
-            .unwrap_or_default();
-        let fresh = self.connect_fresh_dynamic_auto_http3_client(
-            url,
-            origin.clone(),
-            host.to_string(),
-            discovery_start,
-            timeout,
-        );
-        let cached =
-            self.connect_cached_dynamic_auto_http3_client(url, origin, cached_candidates, timeout);
-        race_dynamic_auto_http3_candidates(fresh, cached).await
-    }
-
-    async fn connect_fresh_dynamic_auto_http3_client(
-        &self,
-        url: &Url,
-        origin: String,
-        host: String,
-        discovery_start: std::time::Instant,
-        timeout: TimeoutBudget,
-    ) -> DynamicHttp3ConnectOutcome {
-        let origin_addrs_task =
-            spawn_auto_http3_origin_addrs(self.config.dns_server.clone(), host.clone(), timeout);
-        let records =
-            lookup_auto_http3_https_records(self.config.dns_server.as_deref(), &host, timeout)
-                .await;
-        if let Some(cache) = &self.config.http3_cache {
-            cache.store_https_records(url, self.config.dns_server.as_deref(), &records);
-        }
-        if records.is_empty() {
-            origin_addrs_task.abort();
-            return DynamicHttp3ConnectOutcome::NoCandidates;
-        }
-        let origin_addrs = take_finished_auto_http3_origin_addrs(origin_addrs_task).await;
-        let addrs = auto_http3_addrs_for_records(
-            self.config.dns_server.as_deref(),
-            url,
-            &records,
-            &origin_addrs,
-            timeout,
-        )
-        .await;
-        if addrs.is_empty() {
-            return DynamicHttp3ConnectOutcome::NoCandidates;
-        }
-        record_dns_addrs_trace(&self.config, url, &addrs, discovery_start.elapsed());
-        match self
-            .connect_http3_client_with_addrs(url, origin, addrs, timeout)
-            .await
-        {
-            Ok(result) => DynamicHttp3ConnectOutcome::Connected(result),
-            Err(err) => DynamicHttp3ConnectOutcome::Failed(err),
-        }
-    }
-
-    async fn connect_cached_dynamic_auto_http3_client(
-        &self,
-        url: &Url,
-        origin: String,
-        candidates: Vec<Http3CacheCandidate>,
-        timeout: TimeoutBudget,
-    ) -> DynamicHttp3ConnectOutcome {
-        match self
-            .connect_cached_auto_http3_client_with_candidates(url, origin, candidates, timeout)
-            .await
-        {
-            Some(Ok(result)) => DynamicHttp3ConnectOutcome::Connected(result),
-            Some(Err(err)) => DynamicHttp3ConnectOutcome::Failed(err),
-            None => DynamicHttp3ConnectOutcome::NoCandidates,
-        }
-    }
-
-    async fn connect_auto_tcp_tls(
-        &self,
-        url: &Url,
-        timeout: TimeoutBudget,
-    ) -> Result<AutoTcpConnection, Error> {
-        let trace = connect_direct_tcp_config(&self.config, url, timeout)
-            .await
-            .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-        record_dns_trace(&self.config, url, &trace);
-        let remote_addr = trace.stream.peer_addr().ok();
-        let tcp = trace.tcp_duration;
-        let tls_start = std::time::Instant::now();
-        let (stream, negotiated_h2) = tls_stream_for_config(
-            &self.config,
-            url,
-            Box::pin(trace.stream) as crate::net::DialStream,
-            &alpn_for_config(&self.config),
-            timeout,
-        )
-        .await?;
-        Ok(AutoTcpConnection {
-            stream: PooledStream {
-                inner: stream,
-                negotiated_h2,
-                proxied: false,
-                remote_addr,
-            },
-            negotiated_h2,
-            remote_addr,
-            timing: TransportTiming {
-                tcp: Some(tcp),
-                tls: Some(tls_start.elapsed()),
-                quic: None,
-            },
-        })
-    }
-}
-
-async fn race_primary_fallback<T, StartPrimary, StartFallback, PrimaryFuture, FallbackFuture>(
-    start_primary: StartPrimary,
-    start_fallback: StartFallback,
-    fallback_delay: Duration,
-) -> Result<T, Error>
-where
-    StartPrimary: FnOnce() -> PrimaryFuture,
-    StartFallback: FnOnce() -> FallbackFuture,
-    PrimaryFuture: Future<Output = Result<T, Error>>,
-    FallbackFuture: Future<Output = Result<T, Error>>,
-{
-    let mut primary = Box::pin(start_primary());
-    let mut start_fallback = Some(start_fallback);
-    let mut fallback_task: Option<Pin<Box<FallbackFuture>>> = None;
-    let mut primary_err = None;
-    let mut fallback_err = None;
-    let delay = tokio::time::sleep(fallback_delay);
-    tokio::pin!(delay);
-
-    loop {
-        if primary_err.is_some() && fallback_err.is_some() {
-            return Err(fallback_err
-                .take()
-                .or(primary_err)
-                .expect("at least one race error exists"));
-        }
-
-        if primary_err.is_some() {
-            if fallback_task.is_none() && fallback_err.is_none() {
-                fallback_task = Some(Box::pin(start_fallback
-                    .take()
-                    .expect("fallback has not started yet")(
-                )));
-            }
-            if let Some(task) = fallback_task.as_mut() {
-                match task.as_mut().await {
-                    Ok(winner) => return Ok(winner),
-                    Err(err) => {
-                        fallback_err = Some(err);
-                        fallback_task = None;
-                    }
-                }
-            }
-            continue;
-        }
-
-        if fallback_err.is_some() {
-            match primary.as_mut().await {
-                Ok(winner) => return Ok(winner),
-                Err(err) => primary_err = Some(err),
-            }
-            continue;
-        }
-
-        if fallback_task.is_none() {
-            tokio::select! {
-                result = &mut primary => match result {
-                    Ok(winner) => return Ok(winner),
-                    Err(err) => {
-                        primary_err = Some(err);
-                        fallback_task = Some(Box::pin(
-                            start_fallback
-                                .take()
-                                .expect("fallback has not started yet")(),
-                        ));
-                    }
-                },
-                _ = &mut delay => {
-                    fallback_task = Some(Box::pin(
-                        start_fallback
-                            .take()
-                            .expect("fallback has not started yet")(),
-                    ));
-                }
-            }
-        } else {
-            let fallback = fallback_task.as_mut().expect("fallback task exists");
-            tokio::select! {
-                result = &mut primary => match result {
-                    Ok(winner) => return Ok(winner),
-                    Err(err) => primary_err = Some(err),
-                },
-                result = fallback.as_mut() => match result {
-                    Ok(winner) => return Ok(winner),
-                    Err(err) => {
-                        fallback_err = Some(err);
-                        fallback_task = None;
-                    }
-                },
-            }
-        }
-    }
-}
-
-async fn race_started_auto_connections<H3Future, TcpFuture>(
-    h3: H3Future,
-    tcp: TcpFuture,
-) -> Result<AutoRaceWinner, Error>
-where
-    H3Future: Future<Output = Result<AutoRaceWinner, Error>>,
-    TcpFuture: Future<Output = Result<AutoRaceWinner, Error>>,
-{
-    let mut h3 = Box::pin(h3);
-    let mut tcp = Box::pin(tcp);
-    let mut h3_err = None;
-    let mut tcp_err = None;
-
-    loop {
-        if h3_err.is_some() && tcp_err.is_some() {
-            return Err(tcp_err
-                .take()
-                .or(h3_err)
-                .expect("at least one race error exists"));
-        }
-
-        match (h3_err.is_some(), tcp_err.is_some()) {
-            (false, false) => {
-                tokio::select! {
-                    result = h3.as_mut() => match result {
-                        Ok(winner) => return Ok(winner),
-                        Err(err) => h3_err = Some(err),
-                    },
-                    result = tcp.as_mut() => match result {
-                        Ok(winner) => return Ok(winner),
-                        Err(err) => tcp_err = Some(err),
-                    },
-                }
-            }
-            (true, false) => match tcp.as_mut().await {
-                Ok(winner) => return Ok(winner),
-                Err(err) => tcp_err = Some(err),
-            },
-            (false, true) => match h3.as_mut().await {
-                Ok(winner) => return Ok(winner),
-                Err(err) => h3_err = Some(err),
-            },
-            (true, true) => unreachable!("handled at top of loop"),
-        }
-    }
-}
-
-enum DynamicHttp3ConnectOutcome {
-    Connected(Http3ConnectResult),
-    Failed(Error),
-    NoCandidates,
-}
-
-async fn race_dynamic_auto_http3_candidates<FreshFuture, CachedFuture>(
-    fresh: FreshFuture,
-    cached: CachedFuture,
-) -> Result<Http3ConnectResult, Error>
-where
-    FreshFuture: Future<Output = DynamicHttp3ConnectOutcome>,
-    CachedFuture: Future<Output = DynamicHttp3ConnectOutcome>,
-{
-    let mut fresh = Box::pin(fresh);
-    let prompt_fresh = tokio::task::yield_now();
-    tokio::pin!(prompt_fresh);
-
-    let mut fresh_done = false;
-    let mut cached_done = false;
-    let mut fresh_err = None;
-    let mut cached_err = None;
-
-    tokio::select! {
-        result = fresh.as_mut() => {
-            fresh_done = true;
-            if let Some(result) = record_dynamic_http3_outcome(result, &mut fresh_err) {
-                return Ok(result);
-            }
-        }
-        _ = &mut prompt_fresh => {}
-    }
-
-    let mut cached = Box::pin(cached);
-
-    loop {
-        if fresh_done && cached_done {
-            return Err(fresh_err
-                .or(cached_err)
-                .unwrap_or_else(|| Error::connect("no HTTP/3 candidates discovered")));
-        }
-
-        match (fresh_done, cached_done) {
-            (false, false) => {
-                tokio::select! {
-                    result = fresh.as_mut() => {
-                        fresh_done = true;
-                        if let Some(result) = record_dynamic_http3_outcome(result, &mut fresh_err) {
-                            return Ok(result);
-                        }
-                    }
-                    result = cached.as_mut() => {
-                        cached_done = true;
-                        if let Some(result) = record_dynamic_http3_outcome(result, &mut cached_err) {
-                            return Ok(result);
-                        }
-                    }
-                }
-            }
-            (false, true) => {
-                let result = fresh.as_mut().await;
-                fresh_done = true;
-                if let Some(result) = record_dynamic_http3_outcome(result, &mut fresh_err) {
-                    return Ok(result);
-                }
-            }
-            (true, false) => {
-                let result = cached.as_mut().await;
-                cached_done = true;
-                if let Some(result) = record_dynamic_http3_outcome(result, &mut cached_err) {
-                    return Ok(result);
-                }
-            }
-            (true, true) => unreachable!("handled at top of loop"),
-        }
-    }
-}
-
-fn record_dynamic_http3_outcome(
-    outcome: DynamicHttp3ConnectOutcome,
-    error: &mut Option<Error>,
-) -> Option<Http3ConnectResult> {
-    match outcome {
-        DynamicHttp3ConnectOutcome::Connected(result) => Some(result),
-        DynamicHttp3ConnectOutcome::Failed(err) => {
-            *error = Some(err);
-            None
-        }
-        DynamicHttp3ConnectOutcome::NoCandidates => None,
-    }
-}
-
-fn spawn_auto_http3_origin_addrs(
-    dns_server: Option<String>,
-    host: String,
-    timeout: TimeoutBudget,
-) -> JoinHandle<Vec<SocketAddr>> {
-    tokio::spawn(async move {
-        crate::net::resolve_host(&host, dns_server.as_deref(), timeout)
-            .await
-            .unwrap_or_default()
-    })
-}
-
-async fn take_finished_auto_http3_origin_addrs(
-    handle: JoinHandle<Vec<SocketAddr>>,
-) -> Vec<SocketAddr> {
-    if !handle.is_finished() {
-        tokio::task::yield_now().await;
-    }
-    if handle.is_finished() {
-        handle.await.unwrap_or_default()
-    } else {
-        handle.abort();
-        Vec::new()
-    }
-}
-
-async fn lookup_auto_http3_https_records(
-    dns_server: Option<&str>,
-    host: &str,
-    timeout: TimeoutBudget,
-) -> Vec<SvcbRecord> {
-    let Some(timeout) = auto_http3_lookup_timeout(timeout) else {
-        return Vec::new();
-    };
-    let resolver = dns_server
-        .map(HttpsRecordResolver::Custom)
-        .unwrap_or(HttpsRecordResolver::System);
-    tokio::time::timeout(
-        timeout,
-        crate::dns::svcb::lookup_https_records(resolver, host, Some(timeout)),
-    )
-    .await
-    .ok()
-    .and_then(Result::ok)
-    .unwrap_or_default()
-}
-
-fn auto_http3_lookup_timeout(timeout: TimeoutBudget) -> Option<Duration> {
-    let max_lookup = crate::net::HAPPY_EYEBALLS_FALLBACK_DELAY;
-    match timeout.remaining().ok()? {
-        None => Some(max_lookup),
-        Some(remaining) if remaining <= max_lookup => None,
-        Some(remaining) => Some((remaining - max_lookup).min(max_lookup)),
-    }
-}
-
-async fn auto_http3_addrs_for_records(
-    dns_server: Option<&str>,
-    url: &Url,
-    records: &[SvcbRecord],
-    origin_addrs: &[SocketAddr],
-    timeout: TimeoutBudget,
-) -> Vec<SocketAddr> {
-    let Some(origin_host) = url.host_str() else {
-        return Vec::new();
-    };
-    let Some(origin_port) = url.port_or_known_default() else {
-        return Vec::new();
-    };
-    let mut sorted = records.iter().collect::<Vec<_>>();
-    sorted.sort_by_key(|record| record.priority);
-
-    let mut addrs = Vec::new();
-    for record in sorted {
-        if record.is_alias_mode() || !record.is_usable() || !record.advertises_alpn("h3") {
-            continue;
-        }
-        let port = record.port.unwrap_or(origin_port);
-        let mut record_addrs = auto_http3_hint_addrs(record, origin_addrs, port);
-        if record_addrs.is_empty() {
-            let target = auto_http3_target_host(origin_host, &record.target);
-            if target.eq_ignore_ascii_case(origin_host) && !origin_addrs.is_empty() {
-                record_addrs = origin_addrs
-                    .iter()
-                    .map(|addr| SocketAddr::new(addr.ip(), port))
-                    .collect();
-            } else if target.eq_ignore_ascii_case(origin_host) {
-                record_addrs = crate::net::resolve_host(origin_host, dns_server, timeout)
-                    .await
-                    .unwrap_or_default()
-                    .into_iter()
-                    .map(|addr| SocketAddr::new(addr.ip(), port))
-                    .collect();
-            } else if let Ok(ip) = target.parse::<IpAddr>() {
-                record_addrs.push(SocketAddr::new(ip, port));
-            }
-        }
-        append_unique_socket_addrs(&mut addrs, record_addrs);
-    }
-    addrs
-}
-
-async fn auto_http3_addrs_for_cached_candidates(
-    dns_server: Option<&str>,
-    candidates: &[Http3CacheCandidate],
-    timeout: TimeoutBudget,
-) -> Vec<SocketAddr> {
-    let mut sorted = candidates.iter().collect::<Vec<_>>();
-    sorted.sort_by_key(|candidate| candidate.priority.unwrap_or(u16::MAX));
-    let mut addrs = Vec::new();
-    for candidate in sorted {
-        let record_addrs = if let Ok(ip) = candidate.alt_host.parse::<IpAddr>() {
-            vec![SocketAddr::new(ip, candidate.alt_port)]
-        } else {
-            crate::net::resolve_host(&candidate.alt_host, dns_server, timeout)
-                .await
-                .unwrap_or_default()
-                .into_iter()
-                .map(|addr| SocketAddr::new(addr.ip(), candidate.alt_port))
-                .collect()
-        };
-        append_unique_socket_addrs(&mut addrs, record_addrs);
-    }
-    addrs
-}
-
-fn auto_http3_hint_addrs(
-    record: &SvcbRecord,
-    origin_addrs: &[SocketAddr],
-    port: u16,
-) -> Vec<SocketAddr> {
-    let ipv4 = record
-        .ipv4_hint
-        .iter()
-        .copied()
-        .map(|addr| SocketAddr::new(IpAddr::V4(addr), port))
-        .collect::<Vec<_>>();
-    let ipv6 = record
-        .ipv6_hint
-        .iter()
-        .copied()
-        .map(|addr| SocketAddr::new(IpAddr::V6(addr), port))
-        .collect::<Vec<_>>();
-
-    match origin_addrs.first().map(|addr| addr.ip()) {
-        Some(IpAddr::V4(_)) => crate::net::interleave_socket_addr_families(&ipv4, &ipv6),
-        Some(IpAddr::V6(_)) | None => crate::net::interleave_socket_addr_families(&ipv6, &ipv4),
-    }
-}
-
-fn auto_http3_target_host(origin_host: &str, target: &str) -> String {
-    if target == "." {
-        origin_host.to_string()
-    } else {
-        target.trim_end_matches('.').to_string()
-    }
-}
-
-fn append_unique_socket_addrs(target: &mut Vec<SocketAddr>, addrs: Vec<SocketAddr>) {
-    for addr in addrs {
-        if !target.contains(&addr) {
-            target.push(addr);
-        }
-    }
-}
-
-fn record_dns_addrs_trace(
-    config: &ClientConfig,
-    url: &Url,
-    addrs: &[SocketAddr],
-    duration: Duration,
-) {
-    let Some(resolution) = &config.dns_resolution else {
-        return;
-    };
-    let Some(host) = url.host_str() else {
-        return;
-    };
-    resolution.set(crate::http::client::DnsResolution {
-        socket_addrs: addrs.to_vec(),
-        timing: Some(DnsTiming {
-            host: host.to_string(),
-            addrs: dns_timing_addrs(addrs.iter().map(|addr| addr.ip())),
-            duration,
-        }),
-    });
-}
-
-impl Client {
-    async fn send_http3(
-        &self,
-        method: Method,
-        url: Url,
-        headers: HeaderMap,
-        body: Option<Body>,
-        body_deadline: Option<BodyDeadline>,
-    ) -> Result<Response, Error> {
-        if url.scheme() != "https" {
-            return Err(Error::request(format!(
-                "http3: unsupported protocol scheme: {}",
-                url.scheme()
-            )));
-        }
-        let (req, body) = build_h3_request(method, &url, headers, body)?;
-        let pooled = self.http3_client(&url).await?;
-        self.send_http3_request(url, req, body, body_deadline, pooled)
-            .await
-    }
-
-    async fn send_http3_request(
-        &self,
-        url: Url,
-        req: Request<()>,
-        body: Option<Body>,
-        body_deadline: Option<BodyDeadline>,
-        pooled: H3PooledClient,
-    ) -> Result<Response, Error> {
-        let mut sender = pooled.sender.clone();
-        let stream = match sender.send_request(req).await {
-            Ok(stream) => stream,
-            Err(err) => {
-                self.remove_http3_client(&pooled.origin).await;
-                return Err(Error::with_source(
-                    ErrorKind::Request,
-                    format!("http3 request: {err}"),
-                    err,
-                ));
-            }
-        };
-        let (mut send, mut recv) = stream.split();
-        let upload_deadline = body_deadline.clone();
-        let send_task = H3UploadTask::new(tokio::spawn(async move {
-            match body {
-                Some(body) => send_h3_body(&mut send, body, upload_deadline).await,
-                None => send
-                    .finish()
-                    .await
-                    .map_err(|err| Error::body(format!("http3 request body: {err}"))),
-            }
-        }));
-        let response = recv.recv_response().await.map_err(|err| {
-            Error::with_source(ErrorKind::Request, format!("http3 response: {err}"), err)
-        })?;
-        let upload_task = if send_task.is_finished() {
-            send_task.await?;
-            None
-        } else {
-            Some(send_task)
-        };
-        Ok(Response::from_h3(
-            url,
-            response,
-            recv,
-            sender,
-            upload_task,
-            body_deadline,
-            pooled.remote_addr,
-        ))
-    }
-
-    async fn http3_client(&self, url: &Url) -> Result<H3PooledClient, Error> {
-        let origin = http3_origin(url)?;
-        if let Some(client) = self.h3_pool.lock().await.get(&origin).cloned() {
-            return Ok(client);
-        }
-
-        let result = self.connect_http3_client(url, origin.clone()).await?;
-        if let Some(timing) = &self.config.connection_timing {
-            timing.set(result.timing);
-        }
-        let client = result.client;
-        let mut pool = self.h3_pool.lock().await;
-        Ok(pool.entry(origin).or_insert_with(|| client.clone()).clone())
-    }
-
-    async fn store_http3_client(&self, client: H3PooledClient) {
-        self.h3_pool
-            .lock()
-            .await
-            .insert(client.origin.clone(), client);
-    }
-
-    async fn remove_http3_client(&self, origin: &str) {
-        self.h3_pool.lock().await.remove(origin);
-    }
-
-    async fn connect_auto_http3_client(
-        &self,
-        url: &Url,
-        timeout: TimeoutBudget,
-    ) -> Result<Http3ConnectResult, Error> {
-        let origin = http3_origin(url)?;
-        if let Some(client) = self.h3_pool.lock().await.get(&origin).cloned() {
-            return Ok(Http3ConnectResult {
-                client,
-                timing: TransportTiming::default(),
-            });
-        }
-        let mut fresh_error = None;
-        if let Some(addrs) = self
-            .config
-            .auto_http3
-            .as_ref()
-            .map(|config| config.addrs.clone())
-            .filter(|addrs| !addrs.is_empty())
-        {
-            match self
-                .connect_http3_client_with_addrs(url, origin.clone(), addrs, timeout)
-                .await
-            {
-                Ok(result) => return Ok(result),
-                Err(err) => fresh_error = Some(err),
-            }
-        }
-        let cached_error = match self
-            .connect_cached_auto_http3_client(url, origin, timeout)
-            .await
-        {
-            Some(Ok(result)) => return Ok(result),
-            Some(Err(err)) => Some(err),
-            None => None,
-        };
-        Err(fresh_error
-            .or(cached_error)
-            .unwrap_or_else(|| Error::connect("no HTTP/3 candidates discovered")))
-    }
-
-    async fn connect_cached_auto_http3_client(
-        &self,
-        url: &Url,
-        origin: String,
-        timeout: TimeoutBudget,
-    ) -> Option<Result<Http3ConnectResult, Error>> {
-        let cache = self.config.http3_cache.as_ref()?;
-        let candidates = cache.candidates(url, self.config.dns_server.as_deref());
-        self.connect_cached_auto_http3_client_with_candidates(url, origin, candidates, timeout)
-            .await
-    }
-
-    async fn connect_cached_auto_http3_client_with_candidates(
-        &self,
-        url: &Url,
-        origin: String,
-        candidates: Vec<Http3CacheCandidate>,
-        timeout: TimeoutBudget,
-    ) -> Option<Result<Http3ConnectResult, Error>> {
-        let cache = self.config.http3_cache.as_ref()?;
-        let addrs = auto_http3_addrs_for_cached_candidates(
-            self.config.dns_server.as_deref(),
-            &candidates,
-            timeout,
-        )
-        .await;
-        if addrs.is_empty() {
-            return None;
-        }
-        let result = self
-            .connect_http3_client_with_addrs(url, origin, addrs, timeout)
-            .await;
-        if result.is_err() {
-            cache.remove_candidates(url, self.config.dns_server.as_deref(), &candidates);
-        }
-        Some(result)
-    }
-
-    async fn connect_http3_client(
-        &self,
-        url: &Url,
-        origin: String,
-    ) -> Result<Http3ConnectResult, Error> {
-        let host = url
-            .host_str()
-            .ok_or_else(|| Error::request("URL host is required"))?;
-        let port = url
-            .port_or_known_default()
-            .ok_or_else(|| Error::request("URL port is required"))?;
-        let timeout = TimeoutBudget::new(self.config.connect_timeout);
-        let (addrs, dns_duration) = if let Some(addrs) = self.config.dns_overrides.get(host) {
-            let mut addrs = addrs.clone();
-            for addr in &mut addrs {
-                addr.set_port(port);
-            }
-            (addrs, None)
-        } else {
-            let dns_start = std::time::Instant::now();
-            let addrs = crate::net::resolve_host(host, self.config.dns_server.as_deref(), timeout)
-                .await
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?
-                .into_iter()
-                .map(|mut addr| {
-                    addr.set_port(port);
-                    addr
-                })
-                .collect();
-            (addrs, Some(dns_start.elapsed()))
-        };
-        if let Some(duration) = dns_duration {
-            record_dns_addrs_trace(&self.config, url, &addrs, duration);
-        }
-        self.connect_http3_client_with_addrs(url, origin, addrs, timeout)
-            .await
-    }
-
-    async fn connect_http3_client_with_addrs(
-        &self,
-        url: &Url,
-        origin: String,
-        mut addrs: Vec<SocketAddr>,
-        timeout: TimeoutBudget,
-    ) -> Result<Http3ConnectResult, Error> {
-        let host = url
-            .host_str()
-            .ok_or_else(|| Error::request("URL host is required"))?;
-        let (mut endpoint, family_filter) = http3_client_endpoint(self.config.local_address)?;
-        if let Some(local_ip) = family_filter {
-            addrs.retain(|addr| addr.ip().is_ipv4() == local_ip.is_ipv4());
-        }
-        let mut tls = self
-            .config
-            .tls_config
-            .clone()
-            .unwrap_or_else(default_tls_config);
-        tls.alpn_protocols = vec![b"h3".to_vec()];
-        let client_config = QuicClientConfig::try_from(tls)
-            .map(|config| quinn::ClientConfig::new(Arc::new(config)))
-            .map_err(|err| Error::request(format!("invalid QUIC TLS configuration: {err}")))?;
-        endpoint.set_default_client_config(client_config);
-        let start = std::time::Instant::now();
-        let connection = connect_http3(endpoint, addrs, host.to_string(), timeout)
-            .await
-            .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-        let remote_addr = connection.remote_address();
-        let timing = TransportTiming {
-            tcp: None,
-            tls: None,
-            quic: Some(start.elapsed()),
-        };
-        let h3_connection = h3_quinn::Connection::new(connection);
-        let (mut driver, sender) = h3::client::new(h3_connection)
-            .await
-            .map_err(|err| Error::connect(format!("http3 handshake: {err}")))?;
-        tokio::spawn(async move {
-            let _ = future::poll_fn(|cx| driver.poll_close(cx)).await;
-        });
-        Ok(Http3ConnectResult {
-            client: H3PooledClient {
-                origin,
-                sender,
-                remote_addr,
-            },
-            timing,
-        })
-    }
-}
-
-impl ClientBuilder {
-    pub(crate) fn build(self) -> Result<Client, Error> {
-        let config = Arc::new(self.config);
-        let connector = TransportConnector {
-            config: config.clone(),
-        };
-        let mut builder = HyperClient::builder(TokioExecutor::new());
-        builder.pool_timer(TokioTimer::new());
-        if matches!(config.mode, Some(HttpVersion::Http2)) {
-            builder.http2_only(true);
-        }
-        Ok(Client {
-            config,
-            pooled: builder.build(connector),
-            h3_pool: Arc::new(Mutex::new(HashMap::new())),
-        })
-    }
-
-    pub(crate) fn use_rustls_tls(self) -> Self {
-        self
-    }
-
-    pub(crate) fn no_brotli(self) -> Self {
-        self
-    }
-
-    pub(crate) fn no_gzip(self) -> Self {
-        self
-    }
-
-    pub(crate) fn no_zstd(self) -> Self {
-        self
-    }
-
-    pub(crate) fn http1_only(mut self) -> Self {
-        self.config.mode = Some(HttpVersion::Http1);
-        self
-    }
-
-    pub(crate) fn http2_prior_knowledge(mut self) -> Self {
-        self.config.mode = Some(HttpVersion::Http2);
-        self
-    }
-
-    pub(crate) fn http3_prior_knowledge(mut self) -> Self {
-        self.config.mode = Some(HttpVersion::Http3);
-        self
-    }
-
-    #[cfg(unix)]
-    pub(crate) fn unix_socket(mut self, path: &str) -> Self {
-        self.config.unix_socket = Some(path.to_string());
-        self
-    }
-
-    pub(crate) fn local_address(mut self, addr: IpAddr) -> Self {
-        self.config.local_address = Some(addr);
-        self
-    }
-
-    pub(crate) fn auto_http3(mut self, config: AutoHttp3Config) -> Self {
-        self.config.auto_http3 = Some(config);
-        self
-    }
-
-    pub(crate) fn auto_http3_discovery(mut self) -> Self {
-        self.config.auto_http3_discovery = true;
-        self
-    }
-
-    pub(crate) fn http3_cache(mut self, cache: Arc<Http3Cache>, learn_alt_svc: bool) -> Self {
-        self.config.http3_cache = Some(cache);
-        self.config.learn_alt_svc = learn_alt_svc;
-        self
-    }
-
-    pub(crate) fn dns_server(mut self, server: String) -> Self {
-        self.config.dns_server = Some(server);
-        self
-    }
-
-    pub(crate) fn dns_resolution(
-        mut self,
-        resolution: crate::http::client::DnsResolutionHandle,
-    ) -> Self {
-        self.config.dns_resolution = Some(resolution);
-        self
-    }
-
-    pub(crate) fn resolve_to_addrs(mut self, host: &str, addrs: &[SocketAddr]) -> Self {
-        self.config
-            .dns_overrides
-            .insert(host.to_string(), addrs.to_vec());
-        self
-    }
-
-    pub(crate) fn tls_config(mut self, config: rustls::ClientConfig) -> Self {
-        self.config.tls_config = Some(config);
-        self
-    }
-
-    pub(crate) fn connect_timeout(mut self, timeout: Duration) -> Self {
-        self.config.connect_timeout = Some(timeout);
-        self
-    }
-
-    pub(crate) fn timeout_with_message(
-        mut self,
-        timeout: Duration,
-        timeout_message: impl Into<String>,
-    ) -> Self {
-        self.config.request_timeout = Some(timeout);
-        self.config.request_timeout_message = Some(timeout_message.into());
-        self
-    }
-
-    pub(crate) fn cookie_provider(
-        mut self,
-        session: Arc<crate::session::PersistentCookieStore>,
-    ) -> Self {
-        self.config.session = Some(session);
-        self
-    }
-
-    pub(crate) fn proxy(mut self, proxy: Proxy) -> Self {
-        self.config.proxies.push(proxy);
-        self
-    }
-
-    pub(crate) fn connection_timing(
-        mut self,
-        timing: crate::http::client::ConnectionTiming,
-    ) -> Self {
-        self.config.connection_timing = Some(timing);
-        self
-    }
-
-    pub(crate) fn redirect(self, _policy: redirect::Policy) -> Self {
-        self
-    }
-}
-
-#[derive(Clone)]
-struct TransportConnector {
-    config: Arc<ClientConfig>,
-}
-
-impl Service<Uri> for TransportConnector {
-    type Response = TokioIo<PooledStream>;
-    type Error = Error;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
-
-    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, uri: Uri) -> Self::Future {
-        let config = self.config.clone();
-        Box::pin(async move { connect_pooled(config, uri).await })
-    }
-}
-
-async fn connect_pooled(
-    config: Arc<ClientConfig>,
-    uri: Uri,
-) -> Result<TokioIo<PooledStream>, Error> {
-    let url = Url::parse(&uri.to_string())
-        .map_err(|err| Error::request(format!("invalid request URI: {err}")))?;
-    let proxy = proxy_for_config(&config, &url);
-    let timeout = TimeoutBudget::new(config.connect_timeout);
-    let tcp_start = std::time::Instant::now();
-    let (mut stream, proxied, uses_tcp, remote_addr, tcp_duration) =
-        dial_stream_for_config(&config, &url, proxy.as_ref(), timeout).await?;
-    let mut timing = TransportTiming {
-        tcp: uses_tcp.then_some(tcp_duration.unwrap_or_else(|| tcp_start.elapsed())),
-        tls: None,
-        quic: None,
-    };
-    let mut negotiated_h2 = false;
-
-    if url.scheme() == "https" {
-        let tls_start = std::time::Instant::now();
-        let (tls, h2) =
-            tls_stream_for_config(&config, &url, stream, &alpn_for_config(&config), timeout)
-                .await?;
-        timing.tls = Some(tls_start.elapsed());
-        stream = tls;
-        negotiated_h2 = h2;
-    }
-
-    if let Some(connection_timing) = &config.connection_timing {
-        connection_timing.set(timing);
-    }
-
-    Ok(TokioIo::new(PooledStream {
-        inner: stream,
-        negotiated_h2,
-        proxied,
-        remote_addr,
-    }))
-}
-
-async fn dial_stream_for_config(
-    config: &ClientConfig,
-    url: &Url,
-    proxy: Option<&Proxy>,
-    timeout: TimeoutBudget,
-) -> Result<
-    (
-        crate::net::DialStream,
-        bool,
-        bool,
-        Option<SocketAddr>,
-        Option<Duration>,
-    ),
-    Error,
-> {
-    match proxy {
-        Some(proxy) if proxy.is_http_proxy() && url.scheme() == "http" => {
-            let proxy_url = crate::net::parse_proxy_url(&proxy.url)
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            let stream =
-                crate::net::dial_http_proxy_stream_with_tls(&proxy.url, &proxy_url, timeout, None)
-                    .await
-                    .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            Ok((stream, true, true, None, None))
-        }
-        Some(proxy) if proxy.is_http_proxy() => {
-            let proxy_url = crate::net::parse_proxy_url(&proxy.url)
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            let proxy_authorization = proxy
-                .basic_auth()
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            let stream = crate::net::dial_http_proxy_tunnel(
-                &proxy.url,
-                &proxy_url,
-                url,
-                timeout,
-                None,
-                proxy_authorization,
-            )
-            .await
-            .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            Ok((stream, false, true, None, None))
-        }
-        Some(proxy)
-            if proxy.scheme().is_ok_and(|scheme| scheme == "socks5")
-                && target_override_addrs(config, url).is_some() =>
-        {
-            let proxy_url = crate::net::parse_proxy_url(&proxy.url)
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            let addrs = target_override_addrs(config, url).expect("checked above");
-            dial_socks5_proxy_to_addrs(&proxy_url, addrs, timeout)
-                .await
-                .map(|stream| (stream, false, true, None, None))
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))
-        }
-        Some(proxy) => crate::net::dial_proxy(&proxy.url, url, None, timeout)
-            .await
-            .map(|stream| (stream, false, true, None, None))
-            .map_err(|err| Error::from_fetch(ErrorKind::Connect, err)),
-        None if config.unix_socket.is_some() => {
-            #[cfg(unix)]
-            {
-                let path = config.unix_socket.as_deref().expect("unix socket checked");
-                UnixStream::connect(path)
-                    .await
-                    .map(|stream| {
-                        (
-                            Box::pin(stream) as crate::net::DialStream,
-                            false,
-                            false,
-                            None,
-                            None,
-                        )
-                    })
-                    .map_err(|err| Error::with_source(ErrorKind::Connect, err.to_string(), err))
-            }
-            #[cfg(not(unix))]
-            {
-                Err(Error::connect("--unix is not supported on this platform"))
-            }
-        }
-        None => {
-            let trace = connect_direct_tcp_config(config, url, timeout)
-                .await
-                .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-            record_dns_trace(config, url, &trace);
-            let remote_addr = trace.stream.peer_addr().ok();
-            let tcp_duration = trace.tcp_duration;
-            Ok((
-                Box::pin(trace.stream) as crate::net::DialStream,
-                false,
-                true,
-                remote_addr,
-                Some(tcp_duration),
-            ))
-        }
-    }
-}
-
-async fn connect_direct_tcp_config(
-    config: &ClientConfig,
-    url: &Url,
-    timeout: TimeoutBudget,
-) -> Result<crate::net::TcpConnectTrace, FetchError> {
-    let host = url
-        .host_str()
-        .ok_or_else(|| FetchError::Message("URL host is required".to_string()))?;
-    let port = url
-        .port_or_known_default()
-        .ok_or_else(|| FetchError::Message("URL port is required".to_string()))?;
-    if let Some(addrs) = config.dns_overrides.get(host) {
-        let mut addrs = addrs.clone();
-        for addr in &mut addrs {
-            addr.set_port(port);
-        }
-        let tcp_start = std::time::Instant::now();
-        let stream = timeout
-            .run(crate::net::connect_first(addrs.clone(), timeout))
-            .await?;
-        return Ok(crate::net::TcpConnectTrace {
-            stream,
-            resolved_addrs: addrs,
-            dns_duration: None,
-            tcp_duration: tcp_start.elapsed(),
-        });
-    }
-    crate::net::connect_tcp_traced(url, config.dns_server.as_deref(), timeout).await
-}
-
-fn record_dns_trace(config: &ClientConfig, url: &Url, trace: &crate::net::TcpConnectTrace) {
-    let Some(duration) = trace.dns_duration else {
-        return;
-    };
-    record_dns_addrs_trace(config, url, &trace.resolved_addrs, duration);
-}
-
-fn dns_timing_addrs(addrs: impl IntoIterator<Item = IpAddr>) -> Vec<IpAddr> {
-    let mut unique = Vec::new();
-    for addr in addrs {
-        if !unique.contains(&addr) {
-            unique.push(addr);
-        }
-    }
-    unique
-}
-
-fn proxy_for_config(config: &ClientConfig, url: &Url) -> Option<Proxy> {
-    config
-        .proxies
-        .iter()
-        .find_map(|proxy| proxy.selected_for(url))
-}
-
-fn target_override_addrs(config: &ClientConfig, url: &Url) -> Option<Vec<SocketAddr>> {
-    let host = url.host_str()?;
-    let port = url.port_or_known_default()?;
-    let mut addrs = config.dns_overrides.get(host)?.clone();
-    for addr in &mut addrs {
-        addr.set_port(port);
-    }
-    (!addrs.is_empty()).then_some(addrs)
-}
-
-async fn dial_socks5_proxy_to_addrs(
-    proxy_url: &Url,
-    addrs: Vec<SocketAddr>,
-    timeout: TimeoutBudget,
-) -> Result<crate::net::DialStream, FetchError> {
-    let mut last_err = None;
-    for addr in addrs {
-        match crate::net::dial_socks5_proxy_to_addr(proxy_url, addr, timeout).await {
-            Ok(stream) => return Ok(stream),
-            Err(err) => last_err = Some(err),
-        }
-    }
-    Err(last_err.unwrap_or_else(|| FetchError::Runtime("lookup returned no addresses".to_string())))
-}
-
-fn alpn_for_config(config: &ClientConfig) -> Vec<Vec<u8>> {
-    match config.mode {
-        Some(HttpVersion::Http1) => vec![b"http/1.1".to_vec()],
-        Some(HttpVersion::Http2) => vec![b"h2".to_vec()],
-        Some(HttpVersion::Http3) => vec![b"h3".to_vec()],
-        None => vec![b"h2".to_vec(), b"http/1.1".to_vec()],
-    }
-}
-
-async fn tls_stream_for_config(
-    config: &ClientConfig,
-    url: &Url,
-    stream: crate::net::DialStream,
-    alpn: &[Vec<u8>],
-    timeout: TimeoutBudget,
-) -> Result<(crate::net::DialStream, bool), Error> {
-    let host = url
-        .host_str()
-        .ok_or_else(|| Error::request("URL host is required"))?;
-    let server_name = ServerName::try_from(host.to_string())
-        .map_err(|_| Error::request(format!("invalid server name '{host}'")))?;
-    let mut tls_config = config.tls_config.clone().unwrap_or_else(default_tls_config);
-    tls_config.alpn_protocols = alpn.to_vec();
-    let stream = timeout
-        .run(async {
-            TlsConnector::from(Arc::new(tls_config))
-                .connect(server_name, stream)
-                .await
-                .map_err(|err| FetchError::Runtime(format!("tls: {err}")))
-        })
-        .await
-        .map_err(|err| Error::from_fetch(ErrorKind::Connect, err))?;
-    let negotiated_h2 = {
-        let (_, conn) = stream.get_ref();
-        matches!(conn.alpn_protocol(), Some(b"h2"))
-    };
-    Ok((Box::pin(stream), negotiated_h2))
-}
-
-fn http3_origin(url: &Url) -> Result<String, Error> {
-    let host = url
-        .host_str()
-        .ok_or_else(|| Error::request("URL host is required"))?;
-    let port = url
-        .port_or_known_default()
-        .ok_or_else(|| Error::request("URL port is required"))?;
-    Ok(format!("{}://{}:{}", url.scheme(), host, port))
-}
-
-fn map_pooled_client_error(err: hyper_util::client::legacy::Error) -> Error {
-    let kind = if err.is_connect() {
-        ErrorKind::Connect
-    } else {
-        ErrorKind::Request
-    };
-    let message = err
-        .source()
-        .map(ToString::to_string)
-        .unwrap_or_else(|| err.to_string());
-    Error::with_source(kind, message, err)
-}
-
-struct PooledStream {
-    inner: crate::net::DialStream,
-    negotiated_h2: bool,
-    proxied: bool,
-    remote_addr: Option<SocketAddr>,
-}
-
-impl Connection for PooledStream {
-    fn connected(&self) -> Connected {
-        let mut connected = Connected::new().proxy(self.proxied);
-        if let Some(remote_addr) = self.remote_addr {
-            connected = connected.extra(PeerAddr(remote_addr));
-        }
-        if self.negotiated_h2 {
-            connected.negotiated_h2()
-        } else {
-            connected
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-struct PeerAddr(SocketAddr);
-
-impl tokio::io::AsyncRead for PooledStream {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut tokio::io::ReadBuf<'_>,
-    ) -> Poll<std::io::Result<()>> {
-        self.inner.as_mut().poll_read(cx, buf)
-    }
-}
-
-impl tokio::io::AsyncWrite for PooledStream {
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<std::io::Result<usize>> {
-        self.inner.as_mut().poll_write(cx, buf)
-    }
-
-    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
-        self.inner.as_mut().poll_flush(cx)
-    }
-
-    fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
-        self.inner.as_mut().poll_shutdown(cx)
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct Proxy {
-    url: String,
-    kind: ProxyKind,
-    no_proxy: Option<NoProxy>,
-    basic_auth: Option<HeaderValue>,
-}
-
-#[derive(Clone)]
-enum ProxyKind {
-    All,
-    Http,
-    Https,
-    System(Arc<matcher::Matcher>),
-}
-
-impl Proxy {
-    pub(crate) fn all(proxy: &str) -> Result<Self, Error> {
-        parse_proxy(proxy)?;
-        Ok(Self {
-            url: proxy.to_string(),
-            kind: ProxyKind::All,
-            no_proxy: None,
-            basic_auth: None,
-        })
-    }
-
-    pub(crate) fn http(proxy: &str) -> Result<Self, Error> {
-        parse_proxy(proxy)?;
-        Ok(Self {
-            url: proxy.to_string(),
-            kind: ProxyKind::Http,
-            no_proxy: None,
-            basic_auth: None,
-        })
-    }
-
-    pub(crate) fn https(proxy: &str) -> Result<Self, Error> {
-        parse_proxy(proxy)?;
-        Ok(Self {
-            url: proxy.to_string(),
-            kind: ProxyKind::Https,
-            no_proxy: None,
-            basic_auth: None,
-        })
-    }
-
-    pub(crate) fn system() -> Self {
-        Self {
-            url: String::new(),
-            kind: ProxyKind::System(Arc::new(matcher::Matcher::from_system())),
-            no_proxy: None,
-            basic_auth: None,
-        }
-    }
-
-    pub(crate) fn no_proxy(mut self, no_proxy: NoProxy) -> Self {
-        self.no_proxy = Some(no_proxy);
-        self
-    }
-
-    pub(crate) fn selected_for_url(&self, url: &Url) -> Option<Self> {
-        self.selected_for(url)
-    }
-
-    pub(crate) fn uses_local_target_dns(&self) -> bool {
-        self.scheme().is_ok_and(|scheme| scheme == "socks5")
-    }
-
-    fn applies_to(&self, url: &Url) -> bool {
-        if self.no_proxy.as_ref().is_some_and(|no_proxy| {
-            crate::http::client::no_proxy_matches_url(url, no_proxy.0.as_deref())
-        }) {
-            return false;
-        }
-        match &self.kind {
-            ProxyKind::All => true,
-            ProxyKind::Http => url.scheme() == "http",
-            ProxyKind::Https => url.scheme() == "https",
-            ProxyKind::System(_) => self.system_selected_for(url).is_some(),
-        }
-    }
-
-    fn selected_for(&self, url: &Url) -> Option<Self> {
-        match &self.kind {
-            ProxyKind::System(_) => self.system_selected_for(url),
-            _ => self.applies_to(url).then(|| self.clone()),
-        }
-    }
-
-    fn system_selected_for(&self, url: &Url) -> Option<Self> {
-        let ProxyKind::System(matcher) = &self.kind else {
-            return None;
-        };
-        let uri = url.as_str().parse::<Uri>().ok()?;
-        let intercepted = matcher.intercept(&uri)?;
-        let proxy_url = intercepted.uri().to_string();
-        let mut proxy = Self::all(&proxy_url).ok()?;
-        proxy.basic_auth = intercepted.basic_auth().cloned();
-        Some(proxy)
-    }
-
-    fn is_http_proxy(&self) -> bool {
-        crate::net::parse_proxy_url(&self.url)
-            .map(|url| matches!(url.scheme(), "http" | "https"))
-            .unwrap_or(false)
-    }
-
-    fn scheme(&self) -> Result<String, FetchError> {
-        crate::net::parse_proxy_url(&self.url).map(|url| url.scheme().to_string())
-    }
-
-    fn basic_auth(&self) -> Result<Option<String>, FetchError> {
-        if let Some(auth) = &self.basic_auth {
-            return auth
-                .to_str()
-                .map(|value| Some(value.to_string()))
-                .map_err(|err| FetchError::Message(format!("invalid proxy authorization: {err}")));
-        }
-        let url = crate::net::parse_proxy_url(&self.url)?;
-        crate::net::proxy_basic_auth(&url)
-    }
-}
-
-fn parse_proxy(proxy: &str) -> Result<(), Error> {
-    crate::net::parse_proxy_url(proxy)
-        .map(|_| ())
-        .map_err(|err| Error::from_fetch(ErrorKind::Request, err))
-}
-
-#[derive(Clone)]
-pub(crate) struct NoProxy(Option<String>);
-
-impl NoProxy {
-    pub(crate) fn from_env() -> Self {
-        Self(
-            std::env::var("NO_PROXY")
-                .or_else(|_| std::env::var("no_proxy"))
-                .ok(),
-        )
-    }
-}
-
-pub(crate) mod redirect {
-    #[derive(Clone, Copy)]
-    pub(crate) struct Policy;
-
-    impl Policy {
-        pub(crate) fn none() -> Self {
-            Self
-        }
-    }
-}
-
-pub(crate) struct RequestBuilder {
-    client: Client,
-    method: Method,
-    url: Url,
-    headers: HeaderMap,
-    body: Option<Body>,
-    version: Option<Version>,
-    timeout: Option<Duration>,
-    timeout_message: Option<String>,
-}
-
-impl RequestBuilder {
-    pub(crate) fn headers(mut self, headers: HeaderMap) -> Self {
-        replace_headers(&mut self.headers, headers);
-        self
-    }
-
-    pub(crate) fn body(mut self, body: impl Into<Body>) -> Self {
-        self.body = Some(body.into());
-        self
-    }
-
-    pub(crate) fn version(mut self, version: Version) -> Self {
-        self.version = Some(version);
-        self
-    }
-
-    pub(crate) fn timeout(mut self, timeout: Duration) -> Self {
-        self.timeout = Some(timeout);
-        self.timeout_message
-            .get_or_insert_with(|| request_timeout_message(timeout));
-        self
-    }
-
-    pub(crate) fn timeout_with_message(
-        mut self,
-        timeout: Duration,
-        timeout_message: impl Into<String>,
-    ) -> Self {
-        self.timeout = Some(timeout);
-        self.timeout_message = Some(timeout_message.into());
-        self
-    }
-
-    pub(crate) async fn send(self) -> Result<Response, Error> {
-        self.client.clone().execute(self).await
-    }
-}
-
-pub(crate) struct Response {
-    url: Url,
-    status: http::StatusCode,
-    version: Version,
-    headers: HeaderMap,
-    body: Body,
-    body_deadline: Option<BodyDeadline>,
-    remote_addr: Option<SocketAddr>,
-}
-
-impl Response {
-    #[cfg(test)]
-    pub(crate) fn test(status: http::StatusCode, headers: HeaderMap, body: Body) -> Self {
-        Self {
-            url: Url::parse("http://updates.example/artifact").expect("static test URL is valid"),
-            status,
-            version: Version::HTTP_11,
-            headers,
-            body,
-            body_deadline: None,
-            remote_addr: None,
-        }
-    }
-
-    fn from_hyper(
-        url: Url,
-        response: http::Response<Incoming>,
-        body_deadline: Option<BodyDeadline>,
-    ) -> Self {
-        Self::from_hyper_with_remote(url, response, body_deadline, None)
-    }
-
-    fn from_hyper_with_remote(
-        url: Url,
-        response: http::Response<Incoming>,
-        body_deadline: Option<BodyDeadline>,
-        fallback_remote_addr: Option<SocketAddr>,
-    ) -> Self {
-        let (parts, body) = response.into_parts();
-        let remote_addr = parts
-            .extensions
-            .get::<PeerAddr>()
-            .map(|addr| addr.0)
-            .or(fallback_remote_addr);
-        Self {
-            url,
-            status: parts.status,
-            version: parts.version,
-            headers: parts.headers,
-            body: Body::map_incoming(body),
-            body_deadline,
-            remote_addr,
-        }
-    }
-
-    fn from_h3<S, O>(
-        url: Url,
-        response: http::Response<()>,
-        stream: h3::client::RequestStream<S, Bytes>,
-        sender: h3::client::SendRequest<O, Bytes>,
-        upload_task: Option<H3UploadTask>,
-        body_deadline: Option<BodyDeadline>,
-        remote_addr: SocketAddr,
-    ) -> Self
-    where
-        S: h3::quic::RecvStream + Send + Unpin + 'static,
-        O: h3::quic::OpenStreams<Bytes> + Send + Unpin + 'static,
-    {
-        let (parts, _) = response.into_parts();
-        Self {
-            url,
-            status: parts.status,
-            version: Version::HTTP_3,
-            headers: parts.headers,
-            body: Body::boxed(H3Body {
-                stream,
-                _sender: sender,
-                upload_task,
-                state: H3BodyState::Data,
-            }),
-            body_deadline,
-            remote_addr: Some(remote_addr),
-        }
-    }
-
-    pub(crate) fn status(&self) -> http::StatusCode {
-        self.status
-    }
-
-    pub(crate) fn version(&self) -> Version {
-        self.version
-    }
-
-    pub(crate) fn headers(&self) -> &HeaderMap {
-        &self.headers
-    }
-
-    pub(crate) fn url(&self) -> &Url {
-        &self.url
-    }
-
-    pub(crate) fn content_length(&self) -> Option<u64> {
-        self.headers
-            .get(CONTENT_LENGTH)
-            .and_then(|value| value.to_str().ok())
-            .and_then(|value| value.parse().ok())
-    }
-
-    pub(crate) fn remote_addr(&self) -> Option<SocketAddr> {
-        self.remote_addr
-    }
-
-    pub(crate) async fn chunk(&mut self) -> Result<Option<Bytes>, Error> {
-        loop {
-            let Some(frame) = read_body_frame(&mut self.body, self.body_deadline.as_ref()).await?
-            else {
-                return Ok(None);
-            };
-            if let Ok(data) = frame.into_data()
-                && !data.is_empty()
-            {
-                return Ok(Some(data));
-            }
-        }
-    }
-
-    pub(crate) fn into_body_with_deadline(self) -> (Body, Option<BodyDeadline>) {
-        (self.body, self.body_deadline)
-    }
-}
-
-impl From<Response> for http::Response<Body> {
-    fn from(response: Response) -> Self {
-        let mut builder = http::Response::builder()
-            .status(response.status)
-            .version(response.version);
-        *builder
-            .headers_mut()
-            .expect("response builder has no error before headers") = response.headers;
-        builder
-            .body(response.body)
-            .expect("response parts are valid")
-    }
-}
-
-pub(crate) async fn read_body_frame(
-    body: &mut Body,
-    deadline: Option<&BodyDeadline>,
-) -> Result<Option<Frame<Bytes>>, Error> {
-    let frame = match deadline {
-        Some(deadline) => tokio::time::timeout_at(deadline.deadline, body.frame())
-            .await
-            .map_err(|_| deadline.timeout_error())?,
-        None => body.frame().await,
-    };
-    match frame {
-        Some(Ok(frame)) => Ok(Some(frame)),
-        Some(Err(err)) => Err(err),
-        None => Ok(None),
-    }
-}
-
-struct H3UploadTask {
-    handle: JoinHandle<Result<(), Error>>,
-}
-
-impl H3UploadTask {
-    fn new(handle: JoinHandle<Result<(), Error>>) -> Self {
-        Self { handle }
-    }
-
-    fn is_finished(&self) -> bool {
-        self.handle.is_finished()
-    }
-}
-
-impl Future for H3UploadTask {
-    type Output = Result<(), Error>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        Pin::new(&mut self.handle).poll(cx).map(|result| {
-            result.unwrap_or_else(|err| Err(Error::body(format!("http3 request body task: {err}"))))
-        })
-    }
-}
-
-impl Drop for H3UploadTask {
-    fn drop(&mut self) {
-        self.handle.abort();
-    }
-}
-
-struct H3Body<S, O>
-where
-    O: h3::quic::OpenStreams<Bytes>,
-{
-    stream: h3::client::RequestStream<S, Bytes>,
-    _sender: h3::client::SendRequest<O, Bytes>,
-    upload_task: Option<H3UploadTask>,
-    state: H3BodyState,
-}
-
-enum H3BodyState {
-    Data,
-    Trailers,
-    Done,
-}
-
-impl<S, O> http_body::Body for H3Body<S, O>
-where
-    S: h3::quic::RecvStream + Unpin,
-    O: h3::quic::OpenStreams<Bytes> + Unpin,
-{
-    type Data = Bytes;
-    type Error = Error;
-
-    fn poll_frame(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
-        loop {
-            if let Some(upload_task) = &mut self.upload_task {
-                match Pin::new(upload_task).poll(cx) {
-                    Poll::Ready(Ok(())) => self.upload_task = None,
-                    Poll::Ready(Err(err)) => {
-                        self.upload_task = None;
-                        return Poll::Ready(Some(Err(err)));
-                    }
-                    Poll::Pending => {}
-                }
-            }
-            match self.state {
-                H3BodyState::Data => match futures_util::ready!(self.stream.poll_recv_data(cx)) {
-                    Ok(Some(mut data)) => {
-                        return Poll::Ready(Some(Ok(Frame::data(
-                            data.copy_to_bytes(data.remaining()),
-                        ))));
-                    }
-                    Ok(None) => {
-                        self.state = H3BodyState::Trailers;
-                    }
-                    Err(err) => {
-                        return Poll::Ready(Some(Err(Error::with_source(
-                            ErrorKind::Body,
-                            format!("http3 body: {err}"),
-                            err,
-                        ))));
-                    }
-                },
-                H3BodyState::Trailers => {
-                    match futures_util::ready!(self.stream.poll_recv_trailers(cx)) {
-                        Ok(Some(trailers)) => {
-                            self.state = H3BodyState::Done;
-                            return Poll::Ready(Some(Ok(Frame::trailers(trailers))));
-                        }
-                        Ok(None) => {
-                            self.state = H3BodyState::Done;
-                            return Poll::Ready(None);
-                        }
-                        Err(err) => {
-                            return Poll::Ready(Some(Err(Error::with_source(
-                                ErrorKind::Body,
-                                format!("http3 trailers: {err}"),
-                                err,
-                            ))));
-                        }
-                    }
-                }
-                H3BodyState::Done => return Poll::Ready(None),
-            }
-        }
-    }
-}
-
-async fn send_h3_body<S>(
-    send: &mut h3::client::RequestStream<S, Bytes>,
-    mut body: Body,
-    deadline: Option<BodyDeadline>,
-) -> Result<(), Error>
-where
-    S: h3::quic::SendStream<Bytes> + Unpin,
-{
-    while let Some(frame) = read_body_frame(&mut body, deadline.as_ref()).await? {
-        if let Ok(data) = frame.into_data()
-            && !data.is_empty()
-        {
-            send.send_data(data).await.map_err(|err| {
-                Error::with_source(ErrorKind::Body, format!("http3 request body: {err}"), err)
-            })?;
-        }
-    }
-    send.finish().await.map_err(|err| {
-        Error::with_source(ErrorKind::Body, format!("http3 request body: {err}"), err)
-    })
-}
-
-fn version_for_cli(version: HttpVersion) -> Version {
-    match version {
-        HttpVersion::Http1 => Version::HTTP_11,
-        HttpVersion::Http2 => Version::HTTP_2,
-        HttpVersion::Http3 => Version::HTTP_3,
-    }
-}
-
-pub(crate) fn extract_url_basic_auth(url: &mut Url) -> Option<(String, Option<String>)> {
-    let username = url.username().to_string();
-    let password = url.password().map(str::to_string);
-    if username.is_empty() && password.is_none() {
-        return None;
-    }
-
-    url.set_username("").ok()?;
-    url.set_password(None).ok()?;
-
-    let username = percent_encoding::percent_decode_str(&username)
-        .decode_utf8()
-        .ok()?
-        .into_owned();
-    let password = match password.as_deref() {
-        Some(password) => Some(
-            percent_encoding::percent_decode_str(password)
-                .decode_utf8()
-                .ok()?
-                .into_owned(),
-        ),
-        None => None,
-    };
-
-    Some((username, password))
-}
-
-pub(crate) fn basic_auth_header_value(username: &str, password: Option<&str>) -> HeaderValue {
-    let raw = format!("{}:{}", username, password.unwrap_or_default());
-    let encoded = base64::engine::general_purpose::STANDARD.encode(raw);
-    let mut value =
-        HeaderValue::from_str(&format!("Basic {encoded}")).expect("base64 is a valid header value");
-    value.set_sensitive(true);
-    value
-}
-
-fn replace_headers(target: &mut HeaderMap, headers: HeaderMap) {
-    let mut last_name = None;
-    for (name, value) in headers {
-        match name {
-            Some(name) => {
-                last_name = Some(name.clone());
-                target.insert(name, value);
-            }
-            None => {
-                if let Some(name) = &last_name {
-                    target.append(name.clone(), value);
-                }
-            }
-        }
-    }
-}
-
-fn http3_client_endpoint(
-    local_address: Option<IpAddr>,
-) -> Result<(quinn::Endpoint, Option<IpAddr>), Error> {
-    let local_addr = http3_endpoint_local_addr(local_address);
-    match quinn::Endpoint::client(local_addr) {
-        Ok(endpoint) => Ok((endpoint, local_address)),
-        Err(err) if local_address.is_none() => {
-            let fallback_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::UNSPECIFIED), 0);
-            quinn::Endpoint::client(fallback_addr)
-                .map(|endpoint| (endpoint, Some(IpAddr::V4(Ipv4Addr::UNSPECIFIED))))
-                .map_err(|fallback_err| {
-                    Error::connect(format!(
-                        "failed to bind HTTP/3 client endpoint to {local_addr}: {err}; \
-                         IPv4 fallback {fallback_addr} also failed: {fallback_err}"
-                    ))
-                })
-        }
-        Err(err) => Err(Error::with_source(ErrorKind::Connect, err.to_string(), err)),
-    }
-}
-
-fn http3_endpoint_local_addr(local_address: Option<IpAddr>) -> SocketAddr {
-    SocketAddr::new(
-        local_address.unwrap_or(IpAddr::V6(Ipv6Addr::UNSPECIFIED)),
-        0,
-    )
-}
-
-async fn connect_http3(
-    endpoint: quinn::Endpoint,
-    addrs: Vec<SocketAddr>,
-    host: String,
-    timeout: TimeoutBudget,
-) -> Result<quinn::Connection, FetchError> {
-    if addrs.is_empty() {
-        return Err(FetchError::Runtime(
-            "lookup returned no addresses".to_string(),
-        ));
-    }
-    connect_http3_staggered(endpoint, addrs, host, timeout).await
-}
-
-async fn connect_http3_staggered(
-    endpoint: quinn::Endpoint,
-    addrs: Vec<SocketAddr>,
-    host: String,
-    timeout: TimeoutBudget,
-) -> Result<quinn::Connection, FetchError> {
-    let mut pending = VecDeque::from(addrs);
-    let mut active = FuturesUnordered::new();
-    let mut last_err = None;
-    start_next_http3_connect(&endpoint, &host, timeout, &mut pending, &mut active);
-    let delay = tokio::time::sleep(crate::net::HAPPY_EYEBALLS_FALLBACK_DELAY);
-    tokio::pin!(delay);
-
-    loop {
-        if active.is_empty() {
-            return Err(last_err.unwrap_or_else(|| {
-                FetchError::Runtime("lookup returned no addresses".to_string())
-            }));
-        }
-        if pending.is_empty() {
-            match active.next().await {
-                Some(Ok(connection)) => return Ok(connection),
-                Some(Err(err)) => last_err = Some(err),
-                None => {}
-            }
-            continue;
-        }
-
-        tokio::select! {
-            result = active.next() => match result {
-                Some(Ok(connection)) => return Ok(connection),
-                Some(Err(err)) => {
-                    last_err = Some(err);
-                    start_next_http3_connect(&endpoint, &host, timeout, &mut pending, &mut active);
-                    delay.as_mut().reset(tokio::time::Instant::now() + crate::net::HAPPY_EYEBALLS_FALLBACK_DELAY);
-                }
-                None => {}
-            },
-            _ = &mut delay => {
-                start_next_http3_connect(&endpoint, &host, timeout, &mut pending, &mut active);
-                delay.as_mut().reset(tokio::time::Instant::now() + crate::net::HAPPY_EYEBALLS_FALLBACK_DELAY);
-            }
-        }
-    }
-}
-
-fn start_next_http3_connect(
-    endpoint: &quinn::Endpoint,
-    host: &str,
-    timeout: TimeoutBudget,
-    pending: &mut VecDeque<SocketAddr>,
-    active: &mut FuturesUnordered<Http3ConnectTask>,
-) {
-    if let Some(addr) = pending.pop_front() {
-        active.push(Http3ConnectTask::new(connect_http3_addr(
-            endpoint.clone(),
-            addr,
-            host.to_string(),
-            timeout,
-        )));
-    }
-}
-
-async fn connect_http3_addr(
-    endpoint: quinn::Endpoint,
-    addr: SocketAddr,
-    host: String,
-    timeout: TimeoutBudget,
-) -> Result<quinn::Connection, FetchError> {
-    let connecting = endpoint
-        .connect(addr, &host)
-        .map_err(|err| FetchError::Runtime(format!("http3 connect {addr}: {err}")))?;
-    timeout
-        .run(async {
-            connecting
-                .await
-                .map_err(|err| FetchError::Runtime(format!("http3 connect {addr}: {err}")))
-        })
-        .await
-}
-
-struct Http3ConnectTask {
-    handle: JoinHandle<Result<quinn::Connection, FetchError>>,
-}
-
-impl Http3ConnectTask {
-    fn new(
-        future: impl Future<Output = Result<quinn::Connection, FetchError>> + Send + 'static,
-    ) -> Self {
-        Self {
-            handle: tokio::spawn(future),
-        }
-    }
-}
-
-impl Future for Http3ConnectTask {
-    type Output = Result<quinn::Connection, FetchError>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        Pin::new(&mut self.handle).poll(cx).map(|result| {
-            result.unwrap_or_else(|err| {
-                Err(FetchError::Runtime(format!(
-                    "http3 connect task failed: {err}"
-                )))
-            })
-        })
-    }
-}
-
-impl Drop for Http3ConnectTask {
-    fn drop(&mut self) {
-        self.handle.abort();
-    }
-}
-
-fn build_request<B>(
-    method: Method,
-    uri: Uri,
-    version: Version,
-    headers: HeaderMap,
-    body: B,
-) -> Result<Request<B>, String> {
-    let mut builder = Request::builder().method(method).uri(uri).version(version);
-    *builder
-        .headers_mut()
-        .expect("request builder has no error before headers") = headers;
-    builder.body(body).map_err(|err| err.to_string())
-}
-
-fn build_h3_request(
-    method: Method,
-    url: &Url,
-    headers: HeaderMap,
-    body: Option<Body>,
-) -> Result<(Request<()>, Option<Body>), Error> {
-    match body {
-        Some(body) => {
-            let request = build_request(method, absolute_uri(url)?, Version::HTTP_3, headers, body)
-                .map_err(Error::request)?;
-            let (parts, body) = request.into_parts();
-            Ok((Request::from_parts(parts, ()), Some(body)))
-        }
-        None => {
-            let request = build_request(
-                method,
-                absolute_uri(url)?,
-                Version::HTTP_3,
-                headers,
-                empty_request_body(),
-            )
-            .map_err(Error::request)?;
-            let (parts, _) = request.into_parts();
-            Ok((Request::from_parts(parts, ()), None))
-        }
-    }
-}
-
-fn empty_request_body() -> impl http_body::Body<Data = Bytes, Error = Error> + Send + 'static {
-    Empty::<Bytes>::new().map_err(|err: Infallible| match err {})
-}
-
-fn default_tls_config() -> rustls::ClientConfig {
-    crate::tls::rustls_platform_client_config().expect("default rustls client config is valid")
-}
-
-fn absolute_uri(url: &Url) -> Result<Uri, Error> {
-    url.as_str()
-        .parse::<Uri>()
-        .map_err(|err| Error::request(format!("invalid request URI: {err}")))
-}
-
-fn origin_form_uri(url: &Url) -> Result<Uri, Error> {
-    let path = if url.path().is_empty() {
-        "/"
-    } else {
-        url.path()
-    };
-    let mut out = path.to_string();
-    if let Some(query) = url.query() {
-        out.push('?');
-        out.push_str(query);
-    }
-    out.parse::<Uri>()
-        .map_err(|err| Error::request(format!("invalid request URI: {err}")))
-}
-
-fn apply_host_header(url: &Url, headers: &mut HeaderMap) -> Result<(), FetchError> {
-    if headers.contains_key(HOST) {
-        return Ok(());
-    }
-    let value = crate::net::http_host_header_value(url)?;
-    headers.insert(
-        HOST,
-        HeaderValue::from_str(&value)
-            .map_err(|err| FetchError::Message(format!("invalid host header: {err}")))?,
-    );
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::Arc;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-
-    #[test]
-    fn replace_headers_preserves_duplicate_values() {
-        let mut source = HeaderMap::new();
-        source.append("x-repeat", HeaderValue::from_static("one"));
-        source.append("x-repeat", HeaderValue::from_static("two"));
-        let mut target = HeaderMap::new();
-
-        replace_headers(&mut target, source);
-
-        let values = target
-            .get_all("x-repeat")
-            .iter()
-            .map(|value| value.to_str().unwrap())
-            .collect::<Vec<_>>();
-        assert_eq!(values, ["one", "two"]);
-    }
-
-    #[test]
-    fn replace_headers_replaces_existing_first_value() {
-        let mut source = HeaderMap::new();
-        source.append("authorization", HeaderValue::from_static("Bearer new"));
-        let mut target = HeaderMap::new();
-        target.append("authorization", HeaderValue::from_static("Bearer old"));
-        target.append(
-            "authorization",
-            HeaderValue::from_static("Bearer duplicate"),
-        );
-
-        replace_headers(&mut target, source);
-
-        let values = target
-            .get_all("authorization")
-            .iter()
-            .map(|value| value.to_str().unwrap())
-            .collect::<Vec<_>>();
-        assert_eq!(values, ["Bearer new"]);
-    }
-
-    #[test]
-    fn extract_url_basic_auth_decodes_and_strips_authority() {
-        let mut url = Url::parse("https://user:open%20sesame@example.com/path").unwrap();
-
-        let auth = extract_url_basic_auth(&mut url).unwrap();
-
-        assert_eq!(auth, ("user".to_string(), Some("open sesame".to_string())));
-        assert_eq!(url.as_str(), "https://example.com/path");
-    }
-
-    #[test]
-    fn http3_client_endpoint_defaults_to_dual_stack_bind() {
-        let default_addr = http3_endpoint_local_addr(None);
-        assert_eq!(default_addr.ip(), IpAddr::V6(Ipv6Addr::UNSPECIFIED));
-
-        let explicit = http3_endpoint_local_addr(Some(IpAddr::V4(Ipv4Addr::LOCALHOST)));
-        assert_eq!(explicit.ip(), IpAddr::V4(Ipv4Addr::LOCALHOST));
-    }
-
-    fn https_record(priority: u16, target: &str, alpn: &[&str], port: Option<u16>) -> SvcbRecord {
-        SvcbRecord {
-            priority,
-            target: target.to_string(),
-            alpn: alpn.iter().map(|value| value.to_string()).collect(),
-            no_default_alpn: false,
-            port,
-            ipv4_hint: Vec::new(),
-            ipv6_hint: Vec::new(),
-            mandatory: Vec::new(),
-            unsupported_mandatory: Vec::new(),
-            ttl: Some(60),
-        }
-    }
-
-    #[test]
-    fn dynamic_auto_http3_hints_follow_origin_family_preference_and_interleave() {
-        let mut record = https_record(1, ".", &["h3"], None);
-        record.ipv4_hint = vec!["192.0.2.1".parse().unwrap(), "192.0.2.2".parse().unwrap()];
-        record.ipv6_hint = vec![
-            "2001:db8::1".parse().unwrap(),
-            "2001:db8::2".parse().unwrap(),
-        ];
-
-        let ipv4_first = auto_http3_hint_addrs(
-            &record,
-            &[SocketAddr::new("198.51.100.1".parse().unwrap(), 443)],
-            443,
-        );
-        assert_eq!(
-            ipv4_first,
-            [
-                SocketAddr::new("192.0.2.1".parse().unwrap(), 443),
-                SocketAddr::new("2001:db8::1".parse().unwrap(), 443),
-                SocketAddr::new("192.0.2.2".parse().unwrap(), 443),
-                SocketAddr::new("2001:db8::2".parse().unwrap(), 443),
-            ]
-        );
-
-        let ipv6_first = auto_http3_hint_addrs(
-            &record,
-            &[SocketAddr::new("2001:db8::10".parse().unwrap(), 443)],
-            443,
-        );
-        assert_eq!(
-            ipv6_first,
-            [
-                SocketAddr::new("2001:db8::1".parse().unwrap(), 443),
-                SocketAddr::new("192.0.2.1".parse().unwrap(), 443),
-                SocketAddr::new("2001:db8::2".parse().unwrap(), 443),
-                SocketAddr::new("192.0.2.2".parse().unwrap(), 443),
-            ]
-        );
-    }
-
-    #[test]
-    fn dynamic_auto_http3_hints_default_to_ipv6_preference_without_origin_addrs() {
-        let mut record = https_record(1, ".", &["h3"], None);
-        record.ipv4_hint = vec!["192.0.2.1".parse().unwrap()];
-        record.ipv6_hint = vec!["2001:db8::1".parse().unwrap()];
-
-        let got = auto_http3_hint_addrs(&record, &[], 443);
-
-        assert_eq!(
-            got,
-            [
-                SocketAddr::new("2001:db8::1".parse().unwrap(), 443),
-                SocketAddr::new("192.0.2.1".parse().unwrap(), 443),
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn dynamic_auto_http3_origin_preference_lookup_allows_custom_dns() {
-        let handle = spawn_auto_http3_origin_addrs(
-            Some("https://dns.example/dns-query".to_string()),
-            "192.0.2.1".to_string(),
-            TimeoutBudget::new(None),
-        );
-
-        let got = take_finished_auto_http3_origin_addrs(handle).await;
-
-        assert_eq!(got, [SocketAddr::new("192.0.2.1".parse().unwrap(), 0)]);
-    }
-
-    #[tokio::test]
-    async fn client_reuses_http1_connections() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let peer_addr = listener.local_addr().unwrap();
-        let url = Url::parse(&format!("http://{peer_addr}")).unwrap();
-        let accepted = Arc::new(AtomicUsize::new(0));
-        let requests = Arc::new(AtomicUsize::new(0));
-        let server = {
-            let accepted = accepted.clone();
-            let requests = requests.clone();
-            tokio::spawn(async move {
-                loop {
-                    let Ok((mut stream, _)) = listener.accept().await else {
-                        break;
-                    };
-                    accepted.fetch_add(1, Ordering::SeqCst);
-                    let requests = requests.clone();
-                    tokio::spawn(async move {
-                        while read_http1_headers(&mut stream).await.is_some() {
-                            requests.fetch_add(1, Ordering::SeqCst);
-                            stream
-                                .write_all(
-                                    b"HTTP/1.1 200 OK\r\ncontent-length: 2\r\nconnection: keep-alive\r\n\r\nok",
-                                )
-                                .await
-                                .unwrap();
-                        }
-                    });
-                }
-            })
-        };
-
-        let client = Client::builder().build().unwrap();
-        for path in ["/one", "/two"] {
-            let mut response = client
-                .request(Method::GET, url.join(path).unwrap())
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(response.remote_addr(), Some(peer_addr));
-            while response.chunk().await.unwrap().is_some() {}
-        }
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), async {
-            while requests.load(Ordering::SeqCst) < 2 {
-                tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-            }
-        })
-        .await
-        .unwrap();
-        assert_eq!(accepted.load(Ordering::SeqCst), 1);
-        server.abort();
-    }
-
-    #[tokio::test]
-    async fn client_reuses_http2_connections() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let peer_addr = listener.local_addr().unwrap();
-        let url = Url::parse(&format!("http://{peer_addr}")).unwrap();
-        let accepted = Arc::new(AtomicUsize::new(0));
-        let requests = Arc::new(AtomicUsize::new(0));
-        let server = {
-            let accepted = accepted.clone();
-            let requests = requests.clone();
-            tokio::spawn(async move {
-                loop {
-                    let Ok((stream, _)) = listener.accept().await else {
-                        break;
-                    };
-                    accepted.fetch_add(1, Ordering::SeqCst);
-                    let requests = requests.clone();
-                    tokio::spawn(async move {
-                        let mut connection = h2::server::handshake(stream).await.unwrap();
-                        while let Some(result) = connection.accept().await {
-                            let (request, mut respond) = result.unwrap();
-                            requests.fetch_add(1, Ordering::SeqCst);
-                            tokio::spawn(async move {
-                                let mut body = request.into_body();
-                                while let Some(chunk) = body.data().await {
-                                    chunk.unwrap();
-                                }
-                                let response = http::Response::builder()
-                                    .status(http::StatusCode::OK)
-                                    .body(())
-                                    .unwrap();
-                                let mut stream = respond.send_response(response, false).unwrap();
-                                stream.send_data(Bytes::from_static(b"ok"), true).unwrap();
-                            });
-                        }
-                    });
-                }
-            })
-        };
-
-        let client = Client::builder().http2_prior_knowledge().build().unwrap();
-        for path in ["/one", "/two"] {
-            let mut response = client
-                .request(Method::GET, url.join(path).unwrap())
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(response.remote_addr(), Some(peer_addr));
-            while response.chunk().await.unwrap().is_some() {}
-        }
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), async {
-            while requests.load(Ordering::SeqCst) < 2 {
-                tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-            }
-        })
-        .await
-        .unwrap();
-        assert_eq!(accepted.load(Ordering::SeqCst), 1);
-        server.abort();
-    }
-
-    #[tokio::test]
-    async fn auto_protocol_race_uses_primary_when_it_wins() {
-        let fallback_started = Arc::new(AtomicUsize::new(0));
-        let fallback_started_for_task = fallback_started.clone();
-
-        let result = race_primary_fallback(
-            || async { Ok("h3") },
-            move || {
-                fallback_started_for_task.fetch_add(1, Ordering::SeqCst);
-                async { Ok("tcp") }
-            },
-            std::time::Duration::from_secs(1),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(result, "h3");
-        assert_eq!(fallback_started.load(Ordering::SeqCst), 0);
-    }
-
-    #[tokio::test]
-    async fn auto_protocol_race_uses_fallback_after_delay() {
-        let result = tokio::time::timeout(
-            std::time::Duration::from_millis(200),
-            race_primary_fallback(
-                || async { std::future::pending::<Result<&'static str, Error>>().await },
-                || async { Ok("tcp") },
-                std::time::Duration::from_millis(10),
-            ),
-        )
-        .await
-        .expect("fallback should start after delay")
-        .unwrap();
-
-        assert_eq!(result, "tcp");
-    }
-
-    #[tokio::test]
-    async fn auto_protocol_race_starts_fallback_after_primary_failure() {
-        let result = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            race_primary_fallback(
-                || async { Err(Error::connect("h3 failed")) },
-                || async { Ok("tcp") },
-                std::time::Duration::from_secs(1),
-            ),
-        )
-        .await
-        .expect("fallback should start immediately after primary failure")
-        .unwrap();
-
-        assert_eq!(result, "tcp");
-    }
-
-    #[tokio::test]
-    async fn auto_protocol_race_reports_fallback_error_when_both_fail() {
-        let err = race_primary_fallback(
-            || async { Err::<&'static str, _>(Error::connect("h3 failed")) },
-            || async { Err::<&'static str, _>(Error::connect("tcp failed")) },
-            std::time::Duration::from_secs(1),
-        )
-        .await
-        .unwrap_err();
-
-        assert_eq!(err.to_string(), "tcp failed");
-    }
-
-    #[tokio::test]
-    async fn auto_protocol_race_drops_losing_primary() {
-        struct NotifyOnDrop(Option<tokio::sync::oneshot::Sender<()>>);
-
-        impl Drop for NotifyOnDrop {
-            fn drop(&mut self) {
-                if let Some(tx) = self.0.take() {
-                    let _ = tx.send(());
-                }
-            }
-        }
-
-        let (dropped_tx, dropped_rx) = tokio::sync::oneshot::channel();
-        let result = race_primary_fallback(
-            || async move {
-                let _notify = NotifyOnDrop(Some(dropped_tx));
-                std::future::pending::<Result<&'static str, Error>>().await
-            },
-            || async { Ok("tcp") },
-            std::time::Duration::from_millis(10),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(result, "tcp");
-        tokio::time::timeout(std::time::Duration::from_secs(1), dropped_rx)
-            .await
-            .unwrap()
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn h3_upload_task_aborts_when_dropped() {
-        struct NotifyOnDrop(Option<tokio::sync::oneshot::Sender<()>>);
-
-        impl Drop for NotifyOnDrop {
-            fn drop(&mut self) {
-                if let Some(tx) = self.0.take() {
-                    let _ = tx.send(());
-                }
-            }
-        }
-
-        let (started_tx, started_rx) = tokio::sync::oneshot::channel();
-        let (dropped_tx, dropped_rx) = tokio::sync::oneshot::channel();
-        let task = H3UploadTask::new(tokio::spawn(async move {
-            let _notify = NotifyOnDrop(Some(dropped_tx));
-            let _ = started_tx.send(());
-            std::future::pending::<Result<(), Error>>().await
-        }));
-
-        started_rx.await.unwrap();
-        drop(task);
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), dropped_rx)
-            .await
-            .unwrap()
-            .unwrap();
-    }
-
-    #[test]
-    fn proxy_selection_prefers_scheme_specific_proxy_before_all_proxy() {
-        let config = Client::builder()
-            .proxy(Proxy::http("http://http-proxy.example:8080").unwrap())
-            .proxy(Proxy::all("http://all-proxy.example:8080").unwrap())
-            .config;
-        let url = Url::parse("http://example.com/").unwrap();
-
-        let proxy = proxy_for_config(&config, &url).unwrap();
-
-        assert_eq!(proxy.url, "http://http-proxy.example:8080");
-    }
-
-    #[test]
-    fn plain_http_proxy_auth_is_added_to_request_headers() {
-        let client = Client::builder()
-            .proxy(Proxy::all("http://user:pass@proxy.example:8080").unwrap())
-            .build()
-            .unwrap();
-        let url = Url::parse("http://example.com/").unwrap();
-        let mut headers = HeaderMap::new();
-
-        client
-            .apply_proxy_authorization(&url, &mut headers)
-            .unwrap();
-
-        assert_eq!(
-            headers
-                .get(PROXY_AUTHORIZATION)
-                .and_then(|value| value.to_str().ok()),
-            Some("Basic dXNlcjpwYXNz")
-        );
-    }
-
-    async fn read_http1_headers(stream: &mut tokio::net::TcpStream) -> Option<Vec<u8>> {
-        let mut raw = Vec::new();
-        let mut buf = [0_u8; 1024];
-        loop {
-            let n = stream.read(&mut buf).await.ok()?;
-            if n == 0 {
-                return None;
-            }
-            raw.extend_from_slice(&buf[..n]);
-            if raw.windows(4).any(|window| window == b"\r\n\r\n") {
-                return Some(raw);
-            }
-        }
-    }
-}
